% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{amplitudeMarkers}
\alias{amplitudeMarkers}
\alias{amplitudeMarkers,CogapsResult-method}
\title{compute amplitude markers statistic}
\usage{
amplitudeMarkers(object, threshold = "all", lp = NA)

\S4method{amplitudeMarkers}{CogapsResult}(object, threshold = "all",
  lp = NA)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{threshold}{the type of threshold to be used. The default "all" will
distribute samples into the pattern with the lowest ranking. The "cut" thresholds
by the first sample to have a lower ranking, i.e. better fit to, a pattern.}

\item{lp}{a vector of weights for each pattern to be used for finding
markers. If NA markers for each pattern of the A matrix will be used.}
}
\value{
By default a non-overlapping list of samples associated with each
\code{lp}.
}
\description{
calculate the most associated pattern for each sample
}
\examples{
data(GIST)
am <- amplitudeMarkers(GIST.result)
}
