% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{binaryA}
\alias{binaryA}
\alias{binaryA,CogapsResult-method}
\title{binary heatmap for standardized feature matrix}
\usage{
binaryA(object, threshold = 3)

\S4method{binaryA}{CogapsResult}(object, threshold = 3)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{threshold}{the number of standard deviations above zero
that an element of Amean must be to get a value of 1}
}
\value{
plots a heatmap of the A Matrix
}
\description{
creates a binarized heatmap of the A matrix
in which the value is 1 if the value in Amean is greater than
threshold * Asd and 0 otherwise
}
\examples{
data(GIST)
result <- CoGAPS(GIST.matrix, nIterations=100)
binMatrix <- binaryA(result, threshold=3)
}
