% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryA.R
\name{binaryA}
\alias{binaryA}
\title{Binary Heatmap for Standardized A Matrix}
\usage{
binaryA(Amean, Asd, threshold = 3)
}
\arguments{
\item{Amean}{the mean estimate for the A matrix}

\item{Asd}{the standard deviations on Amean}

\item{threshold}{the number of standard deviations above zero
that an element of Amean must be to get a value of 1}
}
\value{
plots a heatmap of the A Matrix
}
\description{
Binary Heatmap for Standardized A Matrix
}
\details{
creates a binarized heatmap of the A matrix
in which the value is 1 if the value in Amean is greater than
threshold * Asd and 0 otherwise
}
\examples{
data(SimpSim)
binaryA(SimpSim.result$Amean, SimpSim.result$Asd, threshold=3)
}
