% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{buildReport}
\alias{buildReport}
\title{Information About Package Compilation}
\usage{
buildReport()
}
\value{
string containing build report
}
\description{
Information About Package Compilation
}
\details{
returns information about how the package was compiled, i.e. which
 compiler/version was used, which compile time options were enabled, etc...
}
\examples{
CoGAPS::buildReport()
}

