% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{calcCoGAPSStat}
\alias{calcCoGAPSStat}
\alias{calcCoGAPSStat,CogapsResult-method}
\title{calculate gene set statistics}
\usage{
calcCoGAPSStat(object, GStoGenes, numPerm = 500)

\S4method{calcCoGAPSStat}{CogapsResult}(object, GStoGenes, numPerm = 500)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{GStoGenes}{data.frame or list with gene sets}

\item{numPerm}{number of permutations for null}
}
\value{
gene set statistics for each column of A
}
\description{
calculates the gene set statistics for each
column of A using a Z-score from the elements of the A matrix,
the input gene set, and permutation tests
}
\examples{
data('SimpSim')
result <- CoGAPS(SimpSim.D)
calcCoGAPSStat(result, GStoGenes=GSets, numPerm=500)
}
