% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGeneGSStat.R
\name{calcGeneGSStat}
\alias{calcGeneGSStat}
\title{Probability Gene Belongs in Gene Set}
\usage{
calcGeneGSStat(Amean, Asd, GSGenes, numPerm, Pw = rep(1, ncol(Amean)),
  nullGenes = FALSE)
}
\arguments{
\item{Amean}{A matrix mean values}

\item{Asd}{A matrix standard deviations}

\item{GSGenes}{data.frame or list with gene sets}

\item{numPerm}{number of permutations for null}

\item{Pw}{weight on genes}

\item{nullGenes}{logical indicating gene adjustment}
}
\value{
gene similiarity statistic
}
\description{
Probability Gene Belongs in Gene Set
}
\details{
calculates the probability that a gene
listed in a gene set behaves like other genes in the set within
the given data set
}
\examples{
data('SimpSim')
calcGeneGSStat(SimpSim.result$Amean, SimpSim.result$Asd, GSGenes=GSets[[1]],
numPerm=500)
}
