% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{calcZ}
\alias{calcZ}
\alias{calcZ,CogapsResult-method}
\title{compute z-score matrix}
\usage{
calcZ(object, whichMatrix)

\S4method{calcZ}{CogapsResult}(object, whichMatrix)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{whichMatrix}{either "featureLoadings" or "sampleFactors" indicating which
matrix to calculate the z-score for}
}
\value{
matrix of z-scores
}
\description{
calculates the Z-score for each element based on input mean
and standard deviation matrices
}
\examples{
data(GIST)
featureZScore <- calcZ(GIST.result, "featureLoadings")
}
