% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZ.R
\name{calcZ}
\alias{calcZ}
\title{Compute Z-Score Matrix}
\usage{
calcZ(meanMat, sdMat)
}
\arguments{
\item{meanMat}{matrix of mean values}

\item{sdMat}{matrix of standard deviation values}
}
\value{
matrix of z-scores
}
\description{
Compute Z-Score Matrix
}
\details{
calculates the Z-score for each element based on input mean
and standard deviation matrices
}
\examples{
data(SimpSim)
calcZ(SimpSim.result$Amean, SimpSim.result$Asd)
}
