% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{callInternalCoGAPS}
\alias{callInternalCoGAPS}
\title{make correct call to internal CoGAPS dispatch function, CoGAPS could be
called directly, but to avoid any re-entrant behavior this function is called
instead. It is a light wrapper around cogaps_cpp that handles setting
the distributed parameters}
\usage{
callInternalCoGAPS(
  data,
  allParams,
  uncertainty,
  subsetIndices,
  workerID,
  skipInternalSubset = FALSE
)
}
\arguments{
\item{data}{data in a supported format}

\item{allParams}{list of all parameters}

\item{uncertainty}{uncertainty of data in the same format as data}

\item{subsetIndices}{indices of the subset of data to run on}

\item{workerID}{worker ID for parallelization}

\item{skipInternalSubset}{if TRUE, data is already subsetted and only names
should be subsetted}
}
\value{
CogapsResult object
}
\description{
make correct call to internal CoGAPS dispatch function, CoGAPS could be
called directly, but to avoid any re-entrant behavior this function is called
instead. It is a light wrapper around cogaps_cpp that handles setting
the distributed parameters
}
\keyword{internal}
