% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMatchR.R
\name{cellMatchR}
\alias{cellMatchR}
\title{cellMatchR}
\usage{
cellMatchR(Atot, nSets, cnt, minNS = NA, maxNS = NA, ignore.NA = FALSE,
  bySet = FALSE, plotDen = FALSE, ...)
}
\arguments{
\item{Atot}{a matrix containing the total by set estimates of Pmean output from \code{reOrderBySet}}

\item{nSets}{number of parallel sets used to generate \code{Atot}}

\item{cnt}{number of branches at which to cut dendrogram}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{maxNS}{maximum of individual set contributions a cluster must contain}

\item{ignore.NA}{logical indicating whether or not to ignore NAs from potential over dimensionalization. Default is FALSE.}

\item{bySet}{logical indicating whether to return list of matched set solutions from \code{Atot}}

\item{plotDen}{plot}

\item{...}{additional parameters for \code{agnes}}
}
\value{
a matrix of concensus patterns by samples. If \code{bySet=TRUE} then a list of the set contributions to each
concensus pattern is also returned.
}
\description{
cellMatchR
}
