% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{checkDataMatrix}
\alias{checkDataMatrix}
\title{Check that provided data is valid}
\usage{
checkDataMatrix(data, uncertainty, params)
}
\arguments{
\item{data}{data matrix}

\item{uncertainty}{uncertainty matrix}
}
\value{
throws an error if data has problems
}
\description{
Check that provided data is valid
}

