% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkInputs}
\alias{checkInputs}
\title{check that all inputs are valid}
\usage{
checkInputs(data, uncertainty, allParams)
}
\arguments{
\item{data}{data matrix}

\item{uncertainty}{uncertainty matrix, can be null}

\item{allParams}{list of all parameters}
}
\value{
throws an error if inputs are invalid
}
\description{
check that all inputs are valid
}
\keyword{internal}
