% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{checkpointsEnabled}
\alias{checkpointsEnabled}
\title{Check if package was built with checkpoints enabled}
\usage{
checkpointsEnabled()
}
\value{
true/false if checkpoints are enabled
}
\description{
Check if package was built with checkpoints enabled
}
\examples{
CoGAPS::checkpointsEnabled()
}
