% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\docType{methods}
\name{CoGAPS}
\alias{CoGAPS}
\alias{CoGAPS,character,CogapsParams-method}
\alias{CoGAPS,matrix,CogapsParams-method}
\title{CoGAPS Matrix Factorization Algorithm}
\usage{
CoGAPS(data, params = new("CogapsParams"), uncertainty = NULL,
  fixedMatrix = NULL, checkpointFile = NULL, ...)

\S4method{CoGAPS}{matrix,CogapsParams}(data, params = new("CogapsParams"),
  uncertainty = NULL, fixedMatrix = NULL, checkpointFile = NULL, ...)

\S4method{CoGAPS}{character,CogapsParams}(data, params = new("CogapsParams"),
  uncertainty = NULL, fixedMatrix = NULL, checkpointFile = NULL, ...)
}
\arguments{
\item{data}{File name or R object (see details for supported types)}

\item{params}{CogapsParams object}

\item{uncertainty}{uncertainty matrix (same supported types as data)}

\item{fixedMatrix}{data for fixing the values of either the A or P matrix;
used in conjuction with whichMatrixFixed (see CogapsParams)}

\item{checkpointFile}{name of the checkpoint file}

\item{...}{keeps backwards compatibility with arguments from older versions}
}
\value{
CogapsResult object
}
\description{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix
}
\details{
Currently, raw count matrices are the only supported R object. For
 file types CoGAPS supports csv, tsv, and mtx
}
\examples{
data(SimpSim)
params <- new("CogapsParams")
result <- CoGAPS(SimpSim.D, params)
}

