% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{_PACKAGE}
\name{CoGAPS-package}
\alias{CoGAPS-package}
\title{CoGAPS: Coordinated Gene Activity in Pattern Sets}
\description{
CoGAPS implements a Bayesian MCMC matrix factorization algorithm,
GAPS, and links it to gene set statistic methods to infer
biological process activity.  It can be used to perform
sparse matrix factorization on any data, and when this
data represents biomolecules, to do gene set analysis.
\tabular{ll}{
Package: \tab CoGAPS\cr
Type: \tab Package\cr
Version: \tab 2.99.0\cr
Date: \tab 2018-01-24\cr
License: \tab LGPL\cr
}
}
\references{
Fertig EJ, Ding J, Favorov AV, Parmigiani G, Ochs MF.
CoGAPS: an R/C++ package to identify patterns and biological
process activity in transcriptomic data.
Bioinformatics. 2010 Nov 1;26(21):2792-3
}
\author{
Maintainer: Elana J. Fertig \email{ejfertig@jhmi.edu},
            Michael F. Ochs \email{ochsm@tcnj.edu}
}
