% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{CoGapsFromCheckpoint}
\alias{CoGapsFromCheckpoint}
\title{Restart CoGAPS from Checkpoint File}
\usage{
CoGapsFromCheckpoint(D, S, nFactor, nIter, checkpointFile)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}

\item{nFactor}{number of patterns}

\item{nIter}{number of iterations}

\item{checkpointFile}{path to checkpoint file}

\item{path}{path to checkpoint file}
}
\value{
list with A and P matrix estimates
}
\description{
Restart CoGAPS from Checkpoint File
}
\details{
loads the state of a previous CoGAPS run from a file and
 continues the run from that point
}
\examples{
data(SimpSim)
result <- CoGAPS(SimpSim.D, SimpSim.S, nFactor=3, nOutputs=250)
}

