% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R
\docType{class}
\name{CogapsParams-class}
\alias{CogapsParams-class}
\title{CogapsParams}
\description{
Encapsulates all parameters for the CoGAPS algorithm
}
\section{Slots}{

\describe{
\item{\code{nPatterns}}{number of patterns CoGAPS will learn}

\item{\code{nIterations}}{number of iterations for each phase of the algorithm}

\item{\code{alphaA}}{sparsity parameter for feature matrix}

\item{\code{alphaP}}{sparsity parameter for sample matrix}

\item{\code{maxGibbsMassA}}{atomic mass restriction for feature matrix}

\item{\code{maxGibbsMassP}}{atomic mass restriction for sample matrix}

\item{\code{seed}}{random number generator seed}

\item{\code{singleCell}}{is the data single cell?}

\item{\code{distributed}}{either "genome-wide" or "single-cell" indicating which
distributed algorithm should be used}

\item{\code{nSets}}{[distributed parameter] number of sets to break data into}

\item{\code{cut}}{[distributed parameter] number of branches at which to cut
dendrogram used in pattern matching}

\item{\code{minNS}}{[distributed parameter] minimum of individual set contributions
a cluster must contain}

\item{\code{maxNS}}{[distributed parameter] maximum of individual set contributions
a cluster can contain}

\item{\code{explicitSets}}{[distributed parameter] specify subsets by index or name}

\item{\code{samplingAnnotation}}{[distributed parameter] specify categories along
the rows (cols) to use for weighted sampling}

\item{\code{samplingWeight}}{[distributed parameter] weights associated with 
samplingAnnotation}
}}

