% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CogapsParams.R
\name{CogapsParams}
\alias{CogapsParams}
\title{CogapsParams constructor}
\usage{
CogapsParams(...)
}
\arguments{
\item{...}{parameters for the initialization method}
}
\value{
CogapsParams object
}
\description{
create a CogapsParams object
}
\examples{
params <- CogapsParams(nPatterns=10)
params
}
