% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R
\docType{class}
\name{CogapsResult-class}
\alias{CogapsResult-class}
\title{CogapsResult}
\description{
Contains all output from Cogaps run
}
\section{Slots}{

\describe{
\item{\code{sampleStdDev}}{std dev of the sampled P matrices}

\item{\code{featureStdDev}}{std dev of the sampled A matrices}
}}

