% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{convertDataToMatrix}
\alias{convertDataToMatrix}
\title{convert any acceptable data input to a numeric matrix}
\usage{
convertDataToMatrix(data)
}
\arguments{
\item{data}{data input}
}
\value{
data matrix
}
\description{
convert supported R objects containing the data to a
numeric matrix, if data is a file name do nothing. Exits with an error
if data is not a supported type.
}
\keyword{internal}
