% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{corcut}
\alias{corcut}
\title{cluster patterns together}
\usage{
corcut(allPatterns, cut, minNS)
}
\arguments{
\item{allPatterns}{matrix of all patterns across subsets}

\item{cut}{number of branches at which to cut dendrogram}

\item{minNS}{minimum of individual set contributions a cluster must contain}
}
\value{
patterns listed by which cluster they belong to
}
\description{
cluster patterns together
}
\keyword{internal}
