% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CogapsResult.R
\name{createCogapsResult}
\alias{createCogapsResult}
\title{convert list output from c++ code to a CogapsResult object}
\usage{
createCogapsResult(returnList, allParams)
}
\arguments{
\item{returnList}{list from cogaps_cpp}

\item{allParams}{list of all parameters}
}
\value{
CogapsResult object
}
\description{
convert list output from c++ code to a CogapsResult object
}
\keyword{internal}
