% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGWCoGAPSSets.R
\name{createGWCoGAPSSets}
\alias{createGWCoGAPSSets}
\title{createGWCoGAPSSets}
\usage{
createGWCoGAPSSets(D, S, nSets, simulationName, path = "")
}
\arguments{
\item{nSets}{number of sets for parallelization}

\item{path}{character string indicating were to save resulting data objects}

\item{data}{data matrix with unique rownames}

\item{outRDA}{name of output file}

\item{keep}{logical indicating whether or not to save gene set list. Default is TRUE.}
}
\value{
list with randomly generated sets of genes from whole genome data
}
\description{
\code{createGWCoGAPSSets} factors whole genome data into randomly generated sets for indexing;
}
\examples{
\dontrun{
createGWCoGAPSSet(D,nSets=nSets)
}

}
