% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{createSets}
\alias{createSets}
\title{partition genes/samples into subsets}
\usage{
createSets(data, allParams)
}
\arguments{
\item{data}{either file name or matrix}

\item{allParams}{list of all CoGAPS parameters}
}
\value{
list of sorted subsets of either genes or samples
}
\description{
either genes or samples or partitioned depending on the type
of distributed CoGAPS (i.e. genome-wide or single-cell)
}
\keyword{internal}
