\name{D5}
\docType{data}
\alias{D5}
\title{Expression dataset from Colantuoni et al.}
\description{Small gene expression dataset containing a random set of 1813 genes to support the analysis in Fertig et al. (2013) Pattern identification in time course gene expression data with the CoGAPS matrix factorization. Chapter 6 in MF Ochs (ed) Methods in Molecular Biology: Gene Function Analysis, 2nd Edition, Springer, New York.  Data adapted from Colantuoni et al. (2011) Temporal dynamics and genetic control of transcription in the human prefrontal cortex. Nature, 478:519-523 used for Fertig et al. (2013) Pattern identification in time course gene expression data with the CoGAPS matrix factorization.  Please cite Colantuoni et al. 2011 for use of data and Fertig et al. 2013 in support of time course analyses with CoGAPS.
}
\usage{D5}
\format{Matrix of 1813 rows by 269 columns containing gene expression data subset from Colantuoni et al. 2011.}
\keyword{datasets}
