\name{DGS}
\docType{data}
\alias{DGS}
\title{Simulated gene expresssion data.}
\description{Gene expression data simulated from 3 known true patterns (PGS) with activity in two gene sets (gs) specified in the simulated amplitude (AGS).
}
\usage{DGS}
\format{Matrix of 30 rows by 25 columns of simulated expression measurements.}
\keyword{datasets}
