% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{displayBuildReport}
\alias{displayBuildReport}
\title{Display Information About Package Compilation}
\usage{
displayBuildReport()
}
\value{
display builds information
}
\description{
Display Information About Package Compilation
}
\details{
displays information about how the package was compiled, i.e. which
 compiler/version was used, which compile time options were enabled, etc...
}
\examples{
 CoGAPS::displayBuildReport()
}
