% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{distributedCogaps}
\alias{distributedCogaps}
\title{CoGAPS Distributed Matrix Factorization Algorithm}
\usage{
distributedCogaps(data, allParams, uncertainty)
}
\arguments{
\item{data}{File name or R object (see details for supported types)}

\item{allParams}{list of all parameters used in computation}

\item{uncertainty}{uncertainty matrix (same supported types as data)}
}
\value{
list
}
\description{
runs CoGAPS over subsets of the data and stitches the results
back together
}
\details{
For file types CoGAPS supports csv, tsv, and mtx
}
\keyword{internal}
