% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{fromCSV}
\alias{fromCSV}
\alias{fromCSV,character-method}
\title{read CoGAPS Result object from a directory with a set of csvs 
see toCSV}
\usage{
fromCSV(save_location = ".")

\S4method{fromCSV}{character}(save_location = ".")
}
\arguments{
\item{save_location}{directory to read from}
}
\value{
CogapsResult object
}
\description{
save as csv
}
