% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{gapsCat}
\alias{gapsCat}
\title{wrapper around cat}
\usage{
gapsCat(allParams, ...)
}
\arguments{
\item{allParams}{all cogaps parameters}

\item{...}{arguments forwarded to cat}
}
\value{
conditionally print message
}
\description{
cleans up message printing
}
\keyword{internal}
