% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gapsInterPattern.R
\name{gapsInterPattern}
\alias{gapsInterPattern}
\title{\code{gapsInterPattern} calculates statistics for measuring
	the distance between patterns based on genes associated with 
	the patterns}
\usage{
gapsInterPattern(Amean, Asd, sdThreshold = 3)
}

\arguments{
\item{Amean}{Sampled mean value of the amplitude matrix \eqn{{\bf{A}}} from a run of CoGAPS.}
\item{Asd}{Sampled standard deviation of the amplitude matrix \eqn{{\bf{A}}} from a run of CoGAPS.}
\item{sdThreshold}{How many standard deviations a gene's sampled mean needs to be above 0 to be considered significantly expressed in a pattern.}
}

\description{
\code{gapsInterPattern} calculates the overlap in significant genes between patterns according to the set standard deviation threshold from the mean. 
Returns the names of significant genes, a matrix of overlap fractions, and the mean of this matrix.
A warning is issued if a column of \eqn{{\bf{A}}} does not have significant genes at a given threshold.
}


\details{
	This calculates a statistic for comparing different patterns based on the genes associated with them.
}

\value{
  A list containing:
  \item{SignificantGeneNames}{A list in which each element holds a vector of the name of the significant genes in each column of \eqn{{\bf{A}}.}}
  \item{SignificantGeneTotals}{A vector of significant genes in each respective column.}
  \item{SeparationMatrix}{The matrix containing the overlap fractions for the significant genes in each column of \eqn{{\bf{A}}}.}
  \item{InterPatternValue}{The final value of the Interpattern measure.}
}

\examples{
\dontrun{
## Load data
data('SimpSim')

## Run GAPS matrix decomposition
nIter <- 10
results <- gapsTestRun(SimpSim.D, SimpSim.S, nFactor=3,
                   nEquil=nIter, nSample=nIter)
				   
## Execute the InterPattern function
InterPatternStats <- gapsInterPattern(results$Amean, results$Asd, sdThreshold = 3)
}
}

\seealso{\code{\link{CoGAPS}}}
\keyword{misc}


