% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gapsIntraPattern.R
\name{gapsIntraPattern}
\alias{gapsIntraPattern}
\title{\code{gapsIntraPattern} generates statistics for the similarity
	of gene expression vectors within a pattern}
\usage{
gapsIntraPattern(Amean, Asd, DMatrix, sdThreshold = 3)
}

\arguments{
\item{Amean}{Sampled mean value of the amplitude matrix \eqn{{\bf{A}}} from a run of CoGAPS.}
\item{Asd}{Sampled standard deviation of the amplitude matrix \eqn{{\bf{A}}} from a run of CoGAPS.}
\item{DMatrix}{data matrix}
\item{sdThreshold}{How many standard deviations a gene's sampled mean needs to be above 0 to be considered significantly expressed in a pattern.}
}

\description{
\code{gapsIntraPattern}{Obtains correlation matrices for significantly expressed genes in each column of \eqn{{\bf{A}}}. 
Obtains the means of these matrices, then averages those to get a sense of how closely correlated genes in the patterns of CoGAPS are.}
}


\details{
		This calculates a statistic for determining how tight a pattern is based on the genes associated with it and their distribution in the data.

}

\value{
  A list containing:
  \item{CorrelationMatrices}{A list containing the correlation matrices between the significant genes in each column of \eqn{{\bf{A}}.}}
  \item{CorrelationMatrixMeans}{A list containing the means of the Correlation Matrices.}
  \item{IntraPatternValue}{The final value of the Intrapattern measure.}
  
}

\examples{
\dontrun{
## Load data
data('SimpSim')

## Run GAPS matrix decomposition
nIter <- 10
results <- gapsTestRun(SimpSim.D, SimpSim.S, nFactor=3,
                   nEquil=nIter, nSample=nIter)
				   
## Execute the IntraPattern function
IntraPatternStats <- gapsInterPattern(results$Amean, results$Asd, SimpSim.D, sdThreshold = 3)
}

}

\seealso{\code{\link{CoGAPS}}}
\keyword{misc}


