% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{gapsMapRun}
\alias{gapsMapRun}
\title{Backwards Compatibility with v2}
\usage{
gapsMapRun(D, S, FP, ABins = data.frame(), PBins = data.frame(),
  nFactor = 5, simulation_id = "simulation", nEquil = 1000,
  nSample = 1000, nOutR = 1000, output_atomic = FALSE,
  fixedMatrix = "P", fixedBinProbs = FALSE, fixedDomain = "N",
  sampleSnapshots = TRUE, numSnapshots = 100, alphaA = 0.01,
  nMaxA = 1e+05, max_gibbmass_paraA = 100, alphaP = 0.01, nMaxP = 1e+05,
  max_gibbmass_paraP = 100, seed = -1, messages = TRUE)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}

\item{FP}{data.frame with rows giving fixed patterns for P}

\item{...}{v2 style parameters}
}
\value{
list with A and P matrix estimates
}
\description{
Backwards Compatibility with v2
}

