% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GapsParams.R
\docType{class}
\name{GapsParams-class}
\alias{GapsParams-class}
\title{GapsParams}
\description{
Parameters for running CoGAPS
}
\section{Slots}{

\describe{
\item{\code{nFactor}}{number of patterns (basis vectors, metagenes)}

\item{\code{nEquil}}{number of iterations for burn-in}

\item{\code{nSample}}{number of iterations for sampling}

\item{\code{nOutput}}{how often (number of iterations) to print status in R console}

\item{\code{numSnapshots}}{the number of individual samples to capture}

\item{\code{alphaA}}{sparsity parameter for A domain}

\item{\code{alphaP}}{sparsity parameter for P domain}

\item{\code{maxGibbmassA}}{limit truncated normal to max size}

\item{\code{maxGibbmassP}}{limit truncated normal to max size}

\item{\code{seed}}{positive values are kept, negative values will be overwritten
by a seed generated from the current time}

\item{\code{messages}}{display messages during the run}

\item{\code{singleCelLRNASeq}}{indicate if the data is single cell RNA-seq data}

\item{\code{fixedPatterns}}{matrix of fixed values in either A or P matrix}

\item{\code{whichMatrixFixed}}{characted indicating whether A or P contains
the fixed patterns}

\item{\code{checkpointInterval}}{how often (number of seconds) to create a checkpoint file}

\item{\code{nCores}}{max number of cores to run on}
}}

