% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{gapsRun}
\alias{gapsRun}
\title{Backwards Compatibility with v2}
\usage{
gapsRun(D, S, ...)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}

\item{...}{v2 style parameters}
}
\value{
list with A and P matrix estimates
}
\description{
Backwards Compatibility with v2
}

