% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapsRun.R
\name{gapsRunFromCheckpoint}
\alias{gapsRunFromCheckpoint}
\title{Restart Run from Checkpoint File}
\usage{
gapsRunFromCheckpoint(D, S, path)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}

\item{path}{path to checkpoint file}
}
\value{
list containing information about A and P matrix
}
\description{
Restart Run from Checkpoint File
}

