% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{generateSeeds}
\alias{generateSeeds}
\title{Returns a vector of seeds based on the number of chains and initial seed}
\usage{
generateSeeds(chains = 2, seed = -1)
}
\arguments{
\item{chains}{number of parallel runs}

\item{seed}{initial seed. Negative values use time in seconds as seed}
}
\description{
Returns a vector of seeds based on the number of chains and initial seed
}

