% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getCorrelationToMeanPattern}
\alias{getCorrelationToMeanPattern}
\alias{getCorrelationToMeanPattern,CogapsResult-method}
\title{return correlation between each pattern and the cluster mean}
\usage{
getCorrelationToMeanPattern(object)

\S4method{getCorrelationToMeanPattern}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
CogapsParams object
}
\description{
return correlation between each pattern and the cluster mean
}
\examples{
data(GIST)
corrToMeanPattern <- getCorrelationToMeanPattern(GIST.result)
}
