% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getDimNames}
\alias{getDimNames}
\title{extracts gene/sample names from the data}
\usage{
getDimNames(data, allParams)
}
\arguments{
\item{data}{data matrix}

\item{allParams}{list of all parameters}
}
\value{
list of all parameters with added gene names
}
\description{
extracts gene/sample names from the data
}
\keyword{internal}
