% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getFeatureLoadings}
\alias{getFeatureLoadings}
\alias{getFeatureLoadings,CogapsResult-method}
\title{return featureLoadings matrix from CogapsResult object}
\usage{
getFeatureLoadings(object)

\S4method{getFeatureLoadings}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
featureLoadings matrix
}
\description{
return featureLoadings matrix from CogapsResult object
}
\examples{
data(GIST)
getFeatureLoadings(GIST.result)
}
