% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{getGeneNames}
\alias{getGeneNames}
\title{find the gene names in the data}
\usage{
getGeneNames(data, allParams)
}
\arguments{
\item{data}{data set}

\item{allParams}{list of all cogaps parameters}
}
\value{
vector of gene names
}
\description{
find the gene names in the data
}
