% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getMeanChiSq}
\alias{getMeanChiSq}
\alias{getMeanChiSq,CogapsResult-method}
\title{return chi-sq of final matrices}
\usage{
getMeanChiSq(object)

\S4method{getMeanChiSq}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
chi-sq error
}
\description{
return chi-sq of final matrices
}
\examples{
data(SimpSim)
result <- CoGAPS(SimpSim.data, nIterations=100)
getMeanChiSq(result)
}
