% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getNamesFromData}
\alias{getNamesFromData}
\title{extracts gene/sample names from the data}
\usage{
getNamesFromData(data, allParams, geneNames, sampleNames)
}
\arguments{
\item{data}{data matrix}

\item{allParams}{list of all parameters}

\item{geneNames}{vector of names of genes in data}

\item{sampleNames}{vector of names of samples in data}
}
\value{
list of all parameters with added gene names
}
\description{
extracts gene/sample names from the data
}
\keyword{internal}
