% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getOriginalParameters}
\alias{getOriginalParameters}
\alias{getOriginalParameters,CogapsResult-method}
\title{return original parameters used to generate this result}
\usage{
getOriginalParameters(object)

\S4method{getOriginalParameters}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
CogapsParams object
}
\description{
return original parameters used to generate this result
}
\examples{
data(GIST)
params <- getOriginalParameters(GIST.result)
}
