% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CoGapsParams.R
\docType{methods}
\name{getParam}
\alias{getParam}
\alias{getParam,CoGapsParams-method}
\title{get the value of a parameter}
\usage{
getParam(object, whichParam)

\S4method{getParam}{CoGapsParams}(object, whichParam)
}
\arguments{
\item{whichParam}{a string with the name of the requested parameter}

\item{params}{an object of type CoGapsParams}
}
\value{
the value of the parameter
}
\description{
get the value of a parameter
}
\examples{
 params <- new("CoGapsParams")
 getParam(params, "seed")
}

