% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getPatternHallmarks}
\alias{getPatternHallmarks}
\alias{getPatternHallmarks,CogapsResult-method}
\title{generate statistics associating patterns with MSigDB hallmark gene sets}
\usage{
getPatternHallmarks(object)

\S4method{getPatternHallmarks}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
dataframe of hallmark info
}
\description{
generate statistics associating patterns with MSigDB hallmark gene sets
}
