% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getRetinaSubset}
\alias{getRetinaSubset}
\title{get specified number of retina subsets}
\usage{
getRetinaSubset(n = 1)
}
\arguments{
\item{n}{number of subsets to use}
}
\value{
matrix of RNA counts
}
\description{
combines retina subsets from extdata directory
}
