% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getSampleFactors}
\alias{getSampleFactors}
\alias{getSampleFactors,CogapsResult-method}
\title{return sampleFactors matrix from CogapsResult object}
\usage{
getSampleFactors(object)

\S4method{getSampleFactors}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
sampleFactors matrix
}
\description{
return sampleFactors matrix from CogapsResult object
}
\examples{
data(GIST)
sFactors <- getSampleFactors(GIST.result)
}
