% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{getSampleNames}
\alias{getSampleNames}
\title{find the sample names in the data}
\usage{
getSampleNames(data, allParams)
}
\arguments{
\item{data}{data set}

\item{allParams}{list of all cogaps parameters}
}
\value{
vector of sample names
}
\description{
find the sample names in the data
}
