% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getSubsets}
\alias{getSubsets}
\alias{getSubsets,CogapsResult-method}
\title{return the names of the genes (samples) in each subset}
\usage{
getSubsets(object)

\S4method{getSubsets}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
CogapsParams object
}
\description{
return the names of the genes (samples) in each subset
}
\examples{
data(SimpSim)
result <- CoGAPS(SimpSim.data)
getSubsets(result)
}
