% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getUnmatchedPatterns}
\alias{getUnmatchedPatterns}
\alias{getUnmatchedPatterns,CogapsResult-method}
\title{return unmatched patterns from each subset}
\usage{
getUnmatchedPatterns(object)

\S4method{getUnmatchedPatterns}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
CogapsParams object
}
\description{
return unmatched patterns from each subset
}
\examples{
data(SimpSim)
result <- CoGAPS(SimpSim.data, nIterations=100)
getUnmatchedPatterns(result)
}
