% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getValueOrRds}
\alias{getValueOrRds}
\title{get input that might be an RDS file}
\usage{
getValueOrRds(input)
}
\arguments{
\item{input}{some user input}
}
\value{
if input is an RDS file, read it - otherwise return input
}
\description{
get input that might be an RDS file
}
\keyword{internal}
