% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getVersion}
\alias{getVersion}
\alias{getVersion,CogapsResult-method}
\title{return version number used to generate this result}
\usage{
getVersion(object)

\S4method{getVersion}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
version number
}
\description{
return version number used to generate this result
}
\examples{
data(GIST)
result <- CoGAPS(GIST.matrix, nIterations=100)
getVersion(result)
}
