\name{GIST.D}
\docType{data}
\alias{GIST.D}
\title{Sample GIST gene expression data from Ochs et al. (2009).}
\description{Gene expression data from gastrointestinal stromal tumor cell lines treated with Gleevec.
}
\usage{GIST_TS_20084}
\format{Matrix with 1363 genes by 9 samples of mean gene expression data.}
\references{Ochs, M., Rink, L., Tarn, C., Mburu, S., Taguchi, T., Eisenberg, B., and Godwin, A. (2009). Detection of treatment-induced changes in signaling pathways in gastrointestinal stromal tumors using transcriptomic data. Cancer Res, 69(23), 9125-9132.}
\keyword{datasets}
