% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{GIST.GeneSets}
\alias{GIST.GeneSets}
\title{Simulated dataset to quantify gene set membership in the GIST dataset}
\usage{
GIST.GeneSets
}
\description{
Simulated dataset to quantify gene set membership in the GIST dataset
}
