% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{GIST.uncertainty}
\alias{GIST.uncertainty}
\title{GIST gene expression uncertainty matrix from Ochs et al. (2009)}
\description{
GIST gene expression uncertainty matrix from Ochs et al. (2009)
}
