\name{GSets}
\docType{data}
\alias{GSets}
\alias{GSets}
\title{Simulated dataset to quantify gene set membership.}
\description{Simulated gene sets used to generate amplitude matrix in \code{\link{SimpSim.A}} and corresponding data \code{\link{SimpSim.D}}.}
\usage{GSets}
\format{A \code{\link{list}} containing names of genes in two simulated gene sets used to generate the data in \code{\link{SimpSim.D}}.}
\keyword{datasets}
