% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R
\docType{methods}
\name{initialize,CogapsParams-method}
\alias{initialize,CogapsParams-method}
\title{constructor for CogapsParams}
\usage{
\S4method{initialize}{CogapsParams}(.Object, ...)
}
\description{
constructor for CogapsParams
}
