% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{isRdsFile}
\alias{isRdsFile}
\title{checks if file is rds format}
\usage{
isRdsFile(file)
}
\arguments{
\item{file}{path to file}
}
\value{
TRUE if file is .rds, FALSE if not
}
\description{
checks if file is rds format
}
\keyword{internal}
