% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{MANOVA}
\alias{MANOVA}
\alias{MANOVA,matrix,CogapsResult-method}
\title{MANOVA statistical test for patterns between sample groups}
\usage{
MANOVA(interestedVariables, object)

\S4method{MANOVA}{matrix,CogapsResult}(interestedVariables, object)
}
\arguments{
\item{interestedVariables}{study design for manova}

\item{object}{CogapsResult object}
}
\value{
list of manova fit results
}
\description{
MANOVA statistical test--wraps base R manova
}
