\name{ModSim.P.true}
\docType{data}
\alias{ModSim.P.true}
\title{Simulated gene expresssion data.}
\description{Known true patterns used to simulate gene expression data (ModSim.D).
}
\usage{ModSim}
\format{Matrix of 3 rows by 20 columns containing true patterns used to simulate gene expression data.}
\keyword{datasets}
