% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modsimresult}
\alias{modsimresult}
\title{Result of applying CoGAPS on the Toy example.}
\format{
S4 class ‘CogapsResult’ [package “CoGAPS”] with 7 slots.
}
\usage{
data(modsimresult)
}
\description{
Result of applying CoGAPS on the Toy example.
}
\keyword{datasets}
