% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{ncol_helper}
\alias{ncol_helper}
\title{get number of columns from supported file name or matrix}
\usage{
ncol_helper(data)
}
\arguments{
\item{data}{either a file name or a matrix}
}
\value{
number of columns
}
\description{
get number of columns from supported file name or matrix
}
