% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ncolHelper}
\alias{ncolHelper}
\title{get number of columns from supported file name or matrix}
\usage{
ncolHelper(data)
}
\arguments{
\item{data}{either a file name or a matrix}
}
\value{
number of columns
}
\description{
get number of columns from supported file name or matrix
}
\keyword{internal}
