% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{nrow_helper}
\alias{nrow_helper}
\title{get number of rows from supported file name or matrix}
\usage{
nrow_helper(data)
}
\arguments{
\item{data}{either a file name or a matrix}
}
\value{
number of rows
}
\description{
get number of rows from supported file name or matrix
}
