% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{nrowHelper}
\alias{nrowHelper}
\title{get number of rows from supported file name or matrix}
\usage{
nrowHelper(data)
}
\arguments{
\item{data}{either a file name or a matrix}
}
\value{
number of rows
}
\description{
get number of rows from supported file name or matrix
}
\keyword{internal}
