% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R, R/methods-CogapsParams.R
\docType{methods}
\name{parseDirectParams}
\alias{parseDirectParams}
\alias{parseDirectParams,CogapsParams-method}
\alias{parseDirectParams}
\title{parse list of parameters passed directly to CoGAPS}
\usage{
parseDirectParams(object, args)

\S4method{parseDirectParams}{CogapsParams}(object, args)
}
\arguments{
\item{object}{an object of type CogapsParams}

\item{oldArgs}{named list of arguments}
}
\value{
an object of type CogapsParams
}
\description{
parse list of parameters passed directly to CoGAPS
}
