% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R
\docType{methods}
\name{parseOldParams}
\alias{parseOldParams}
\alias{parseOldParams,CogapsParams-method}
\title{parse list of old-style parameters, store relevant values}
\usage{
parseOldParams(object, oldArgs)

\S4method{parseOldParams}{CogapsParams}(object, oldArgs)
}
\arguments{
\item{object}{an object of type CogapsParams}

\item{oldArgs}{named list of deprecated arguments}
}
\value{
an object of type CogapsParams
}
\description{
parse list of old-style parameters, store relevant values
}

