% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{patternMarkers}
\alias{patternMarkers}
\alias{patternMarkers,CogapsResult-method}
\title{compute pattern markers statistic}
\usage{
patternMarkers(object, threshold = "all", lp = NULL)

\S4method{patternMarkers}{CogapsResult}(object, threshold = "all", lp = NULL)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{threshold}{the type of threshold to be used. The default "all" will
distribute genes into patterns with the lowest ranking as ranked by the
increasing Euclidian distance between gene and the \code{lp}.}

\item{lp}{list of vectors of weights for each pattern to be used for finding
markers. If NULL, list of synthetic one-hot markers for each column of the
featureLoadings matrix will be generated and matched against.}
}
\value{
By default a non-overlapping list of genes associated with each
\code{lp}.

By default a non-overlapping list of genes associated with each \code{lp}. If \code{full=TRUE} a data.frame of
genes rankings with a column for each \code{lp} will also be returned.
}
\description{
calculate the most associated pattern for each gene
}
\examples{
data(GIST)
pm <- patternMarkers(GIST.result)
}
