% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternMarkers.R
\name{patternMarkers}
\alias{patternMarkers}
\title{patternMarkers}
\usage{
patternMarkers(Amatrix = AP$Amean, scaledPmatrix = FALSE, Pmatrix = NA,
  threshold = "cut", lp = NA, full = FALSE, ...)
}
\arguments{
\item{Amatrix}{A matrix of genes by weights resulting from CoGAPS or other NMF decomposition}

\item{scaledPmatrix}{logical indicating whether the corresponding pattern matrix was fixed to have max 1 during decomposition}

\item{Pmatrix}{the corresponding Pmatrix (patterns X samples) for the provided Amatrix (genes x patterns). This must be supplied if scaledPmatrix is FALSE.}

\item{threshold}{the type of threshold to be used. The default "cut" will thresholding by the first gene to have a lower ranking, i.e. better fit to, a pattern. Alternatively, threshold="all" will return all of the genes in rank order for each pattern.}

\item{lp}{a vector of weights for each pattern to be used for finding markers. If NA markers for each pattern of the A matrix will be used.}

\item{full}{logical indicating whether to return the ranks of each gene for each pattern}

\item{...}{}
}
\value{
By default a non-overlapping list of genes associated with each \code{lp}. If \code{full=TRUE} a data.frame of
genes rankings with a column for each \code{lp} will also be returned.
}
\description{
patternMarkers
}
\examples{
\dontrun{
patternMarkers(Amatrix=AP$Amean,scaledPmatrix=FALSE,Pmatrix=NA,threshold="cut")
}

}

