% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{patternMatch}
\alias{patternMatch}
\title{Match Patterns Across Multiple Runs}
\usage{
patternMatch(allPatterns, gapsParams)
}
\arguments{
\item{allPatterns}{matrix of patterns stored in the columns}

\item{gapsParams}{CoGAPS parameters object}
}
\value{
a matrix of consensus patterns
}
\description{
Match Patterns Across Multiple Runs
}
\keyword{internal}
