% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternMatch4Parallel.R
\name{patternMatch4Parallel}
\alias{patternMatch4Parallel}
\title{patternMatch4Parallel}
\usage{
patternMatch4Parallel(Ptot = BySet$P, nSets = nSets, cnt = Cut,
  minNS = minNS, cluster.method = "complete", ignore.NA = FALSE,
  bySet = FALSE, ...)
}
\arguments{
\item{Ptot}{a matrix containing the total by set estimates of Pmean output from \code{reOrderBySet}}

\item{nSets}{number of parallel sets used to generate \code{Ptot}}

\item{cnt}{number of branches at which to cut dendrogram}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{cluster.method}{the agglomeration method to be used for clustering}

\item{ignore.NA}{logical indicating whether or not to ignore NAs from potential over dimensionalization. Default is FALSE.}

\item{bySet}{logical indicating whether to return list of matched set solutions from \code{Ptot}}

\item{...}{}
}
\value{
a matrix of concensus patterns by samples. If \code{bySet=TRUE} then a list of the set contributions to each
concensus pattern is also returned.
}
\description{
patternMatch4Parallel
}
\examples{
dontrun{
matchedPs<-PatternsMatchFN(Ptot=PBySet,nP=nFactor,nS=nSets,cnt=nFactor,minNS=minNS,bySet=FALSE)
}


}
\seealso{
\code{\link{hclust}}
}

