% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternMatch4singleCell.R
\name{patternMatch4singleCell}
\alias{patternMatch4singleCell}
\title{patternMatch4singleCell}
\usage{
patternMatch4singleCell(Atot, nSets, cnt, minNS, cluster.method = "complete",
  ignore.NA = FALSE, bySet = FALSE, ...)
}
\arguments{
\item{Atot}{a matrix containing the total by set estimates of Amean output from \code{reOrderBySet}}

\item{nSets}{number of parallel sets used to generate \code{Atot}}

\item{cnt}{number of branches at which to cut dendrogram}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{cluster.method}{the agglomeration method to be used for clustering}

\item{ignore.NA}{logical indicating whether or not to ignore NAs from potential over dimensionalization. Default is FALSE.}

\item{bySet}{logical indicating whether to return list of matched set solutions from \code{Atot}}

\item{...}{additional parameters for \code{agnes}}
}
\value{
a matrix of concensus patterns by samples. If \code{bySet=TRUE} then a list of the set contributions to each
concensus pattern is also returned.
}
\description{
patternMatch4singleCell
}
\seealso{
\code{\link{agnes}}
}

