% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAtoms.R
\name{plotAtoms}
\alias{plotAtoms}
\title{Plot Number of Atoms}
\usage{
plotAtoms(gapsRes, type = "sampA")
}
\arguments{
\item{gapsRes}{the list resulting from applying GAPS}

\item{type}{the atoms to plot, values are "sampA", "sampP" ,
"equilA", or "equilP" to plot sampling or equilibration teop
atom numbers}
}
\value{
plot
}
\description{
Plot Number of Atoms
}
\details{
a simple plot of the number of atoms
from one of the vectors returned with atom numbers
}
\examples{
data(SimpSim)
plotAtoms(SimpSim.result, type="sampA")
}
