% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiag.R
\name{plotDiag}
\alias{plotDiag}
\title{Diagnostic Plots}
\usage{
plotDiag(gapsRes)
}
\arguments{
\item{gapsRes}{list returned by CoGAPS}
}
\value{
plot
}
\description{
Diagnostic Plots
}
\details{
plots a series of diagnostic plots
}
\examples{
data(SimpSim)
plotDiag(SimpSim.result)
}
