% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{plotPatternHallmarks}
\alias{plotPatternHallmarks}
\alias{plotPatternHallmarks,CogapsResult,list,numeric-method}
\title{generate a barchart of most significant hallmark sets for a pattern}
\usage{
plotPatternHallmarks(object, patternhallmarks, whichpattern = 1)

\S4method{plotPatternHallmarks}{CogapsResult,list,numeric}(object, patternhallmarks, whichpattern = 1)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{patternhallmarks}{output from getPatternHallmarks}

\item{whichpattern}{which pattern to generate bar chart for}
}
\value{
image object of barchart
}
\description{
generate a barchart of most significant hallmark sets for a pattern
}
