% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPatternMarkers.R
\name{plotPatternMarkers}
\alias{plotPatternMarkers}
\title{plotPatternMarkers}
\usage{
plotPatternMarkers(data = NA, patternMarkers = NA, patternPalette = NA,
  sampleNames = NA, samplePalette = NULL, colDenogram = TRUE,
  heatmapCol = "bluered", scale = "row", ...)
}
\arguments{
\item{data}{the dataset from which the patterns where generated}

\item{patternMarkers}{the list of genes generated from the patternMarkers function}

\item{patternPalette}{a vector indicating what color should be used for each pattern}

\item{sampleNames}{names of the samples to use for labeling}

\item{samplePalette}{a vector indicating what color should be used for each sample}

\item{colDenogram}{logical indicating whether to display sample denogram}

\item{heatmapCol}{pallelet giving color scheme for heatmap}

\item{scale}{character indicating if the values should be centered and scaled in either 
the row direction or the column direction, or none. The default is "row".}

\item{...}{additional graphical parameters to be passed to \code{heatmap.2}}
}
\value{
heatmap of the \code{data} values for the \code{patternMarkers}
}
\description{
plotPatternMarkers
}
\examples{
\dontrun{
plotPatternMarkers(data=p,patternMarkers=PatternMarkers,patternPalette=NULL,sampleNames=pd$sample,
samplePalette=pd$color,colDenogram=TRUE,heatmapCol="bluered", scale='row')
}
}
\seealso{
\code{\link{heatmap.2}}
}

