% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{plotResiduals}
\alias{plotResiduals}
\alias{plotResiduals,CogapsResult-method}
\alias{plotResiduals}
\title{plot of residuals}
\usage{
plotResiduals(object, data, uncertainty = NULL)

\S4method{plotResiduals}{CogapsResult}(object, data, uncertainty = NULL)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{data}{original data matrix run through GAPS}

\item{uncertainty}{original standard deviation matrix run through GAPS}
}
\value{
creates a residual plot
}
\description{
calculate residuals and produce heatmap
}
\examples{
data(SimpSim)
result <- CoGAPS(SimpSim.D)
plotResiduals(result, SimpSim.D)
}
