% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postFixed4Parallel.R
\name{postFixed4Parallel}
\alias{postFixed4Parallel}
\title{Post Processing of Parallel Output}
\usage{
postFixed4Parallel(AP.fixed, setValues, setMatrix = "P")
}
\arguments{
\item{AP.fixed}{output of parallel gapsMapRun calls with same FP}

\item{setValues}{data.frame with rows giving fixed patterns for P used as input
for gapsMapRun}

\item{setMatrix}{which matrix, A or P}
}
\value{
list of two data.frames containing the A matrix estimates or their
corresponding standard deviations from output of parallel CoGAPS
}
\description{
Post Processing of Parallel Output
}
