% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postFixed4SC.R
\name{postFixed4SC}
\alias{postFixed4SC}
\title{Post Processing of Parallel Output}
\usage{
postFixed4SC(AP.fixed, setAs)
}
\arguments{
\item{AP.fixed}{output of parallel gapsMapRun calls with same FP}

\item{setAs}{data.frame with rows giving fixed patterns for A used as input
for gapsMapRun}
}
\value{
list of two data.frames containing the A matrix estimates or their
corresponding standard deviations from output of parallel CoGAPS
}
\description{
Post Processing of Parallel Output
}
