% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{reconstructGene}
\alias{reconstructGene}
\alias{reconstructGene,CogapsResult-method}
\title{reconstruct gene}
\usage{
reconstructGene(object, genes = NULL)

\S4method{reconstructGene}{CogapsResult}(object, genes = NULL)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{genes}{an index of the gene or genes of interest}
}
\value{
the D' estimate of a gene or set of genes
}
\description{
reconstruct gene
}
\examples{
data(GIST)
estimatedD <- reconstructGene(GIST.result)
}
