% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructGene.R
\name{reconstructGene}
\alias{reconstructGene}
\title{Reconstruct Gene}
\usage{
reconstructGene(A, P, genes = NA)
}
\arguments{
\item{A}{A matrix estimates}

\item{P}{corresponding P matrix estimates}

\item{genes}{an index of the gene or genes of interest}
}
\value{
the D' estimate of a gene or set of genes
}
\description{
Reconstruct Gene
}
\examples{
data(SimpSim)
reconstructGene(SimpSim.result$Amean, SimpSim.result$Pmean)
}
