% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderByPatternMatch.R
\name{reorderByPatternMatch}
\alias{reorderByPatternMatch}
\title{Reorder By Pattern Match}
\usage{
reorderByPatternMatch(P, matchTo)
}
\arguments{
\item{P}{matrix to be matched}

\item{matchTo}{matrix to match P to}
}
\value{
matched patterns
}
\description{
Reorder By Pattern Match
}
