% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reOrderBySet.R
\name{reOrderBySet}
\alias{reOrderBySet}
\title{reOrderBySet}
\usage{
reOrderBySet(AP, nFactor, nSets)
}
\arguments{
\item{AP}{output of GWCoGAPS in parallel}

\item{nFactor}{number of patterns}

\item{nSets}{number of sets}
}
\value{
a list containing the \code{nSets} sets solution for Amean under "A",
Pmean under "P", and Asd under "Asd"
}
\description{
reOrderBySet
}
\details{
restructures output of CoGAPS into a list containing each sets
solution for Amean, Pmean, and Asd
}

