% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{Plot of Residuals}
\usage{
residuals(AMean_Mat, PMean_Mat, D, S)
}
\arguments{
\item{D}{original data matrix run through GAPS}

\item{S}{original standard deviation matrix run through GAPS}

\item{Amean}{matrix of mean values for A from GAPS}

\item{Pmean}{matrix of mean values for P from GAPS}
}
\description{
Plot of Residuals
}
\details{
calculate residuals and produce heatmap
}

