% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{sampleUniformly}
\alias{sampleUniformly}
\title{subset data by uniformly partioning rows (cols)}
\usage{
sampleUniformly(allParams, total, setSize)
}
\arguments{
\item{allParams}{list of all CoGAPS parameters}

\item{total}{total number of rows (cols) that are being paritioned}

\item{setSize}{the size of each subset of the total}
}
\value{
list of subsets
}
\description{
subset data by uniformly partioning rows (cols)
}
\keyword{internal}
