% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{sampleWithAnnotationWeights}
\alias{sampleWithAnnotationWeights}
\title{subset rows (cols) proportional to the user provided weights}
\usage{
sampleWithAnnotationWeights(allParams, setSize)
}
\arguments{
\item{allParams}{list of all CoGAPS parameters}

\item{setSize}{the size of each subset of the total}
}
\value{
list of subsets
}
\description{
subset rows (cols) proportional to the user provided weights
}
\keyword{internal}
