% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{sampleWithExplictSets}
\alias{sampleWithExplictSets}
\title{use user provided subsets}
\usage{
sampleWithExplictSets(allParams)
}
\arguments{
\item{allParams}{list of all CoGAPS parameters}

}
\value{
list of subsets
}
\description{
use user provided subsets
}
\keyword{internal}
