% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCoGAPS.R
\name{scCoGapsFromCheckpoint}
\alias{scCoGapsFromCheckpoint}
\title{Restart a scCoGAPS run from a Checkpoint}
\usage{
scCoGapsFromCheckpoint(simulationName, nCores, cut = NA, minNS = NA, ...)
}
\arguments{
\item{simulationName}{name of this simulation}

\item{nCores}{number of cores for parallelization. If left to the default NA, nCores = nSets.}

\item{cut}{number of branches at which to cut dendrogram used in patternMatch4Parallel}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{...}{additional parameters to be fed into \code{gapsRun} and \code{gapsMapRun}}
}
\value{
list of A and P estimates
}
\description{
Restart a scCoGAPS run from a Checkpoint
}
\examples{
data(SimpSim)
sim_name <- "example"
createscCoGAPSSets(SimpSim.D, nSets=2, simulationName=sim_name)
trash <- scCoGAPS(sim_name, nFactor=3, nEquil=200, nSample=200)
scCoGapsFromCheckpoint(sim_name, 2)
}

