% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R, R/methods-CogapsParams.R
\docType{methods}
\name{setAnnotationWeights}
\alias{setAnnotationWeights}
\alias{setAnnotationWeights,CogapsParams-method}
\title{set the annotation labels and weights for subsetting the data}
\usage{
setAnnotationWeights(object, annotation, weights)

\S4method{setAnnotationWeights}{CogapsParams}(object, annotation, weights)
}
\arguments{
\item{object}{an object of type CogapsParams}

\item{annotation}{vector of labels}

\item{weights}{vector of weights}
}
\value{
the modified params object
}
\description{
these parameters  are interrelated so they must be set together
}
\examples{
 params <- new("CogapsParams")
 params <- setAnnotationWeights(params, c('a', 'b', 'c'), c(1,2,1))
}
