% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R, R/methods-CogapsParams.R
\docType{methods}
\name{setDistributedParams}
\alias{setDistributedParams}
\alias{setDistributedParams,CogapsParams-method}
\alias{setDistributedParams}
\title{set the value of parameters for distributed CoGAPS}
\usage{
setDistributedParams(object, cut = NULL, minNS = NULL, maxNS = NULL)

\S4method{setDistributedParams}{CogapsParams}(object, cut = NULL,
  minNS = NULL, maxNS = NULL)
}
\arguments{
\item{object}{an object of type CogapsParams}

\item{cut}{a distributed CoGAPS parameter}

\item{minNS}{a distributed CoGAPS parameter}

\item{maxNS}{a distributed CoGAPS parameter}
}
\value{
the modified params object
}
\description{
these parameters  are interrelated so they must be set together
}
\examples{
 params <- new("CogapsParams")
 params <- setDistributedParams(3, 2, 4)
}
