% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CoGapsParams.R
\docType{methods}
\name{setParam}
\alias{setParam}
\alias{setParam,CoGapsParams-method}
\title{set the value of a parameter}
\usage{
setParam(object, whichParam, value)

\S4method{setParam}{CoGapsParams}(object, whichParam, value)
}
\arguments{
\item{whichParam}{a string with the name of the parameter to be changed}

\item{value}{the value to set the parameter to}

\item{params}{an object of type CoGapsParams}
}
\value{
the modified params object
}
\description{
set the value of a parameter
}
\examples{
 params <- new("CoGapsParams")
 params <- setParam(params, "seed", 123)
}

