\name{SimpSim.A}
\docType{data}
\alias{SimpSim.A}
\title{Simulated data}
\description{True amplitude matrix generated from gene sets in \code{\link{GSets}} used to generate simulated data in \code{\link{SimpSim.D}}.}
\usage{SimpSim.A}
\format{Matrix with 30 genes by 3 patterns of true amplitude used to generate simulated data.}
\keyword{datasets}
