\name{SimpSim.D}
\docType{data}
\alias{SimpSim.D}
\title{Simulated data}
\description{Simulated gene expression data from true patterns in \code{\link{SimpSim.P}} and amplitude in \code{\link{SimpSim.A}}.}
\usage{SimpSim.D}
\format{Matrix with 30 genes by 20 samples of simulated gene expression data.}
\keyword{datasets}
