\name{SimpSim.P}
\docType{data}
\alias{SimpSim.P}
\title{Simulated data}
\description{True pattern matrix used to generate simulated data in \code{\link{SimpSim.D}}.}
\usage{SimpSim.P}
\format{Matrix with 3 patterns by 20 samples of true patterns used to generate simulated data.}
\keyword{datasets}
