\name{SimpSim.S}
\docType{data}
\alias{SimpSim.S}
\title{Simulated data}
\description{Standard deviation of simulated gene expression data from true patterns in \code{\link{SimpSim.P}} and amplitude in \code{\link{SimpSim.A}}.}
\usage{SimpSim.S}
\format{Matrix with 30 genes by 20 samples of containing standard deviation of simulated gene expression data.}
\keyword{datasets}
