% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{startupMessage}
\alias{startupMessage}
\title{write start up message}
\usage{
startupMessage(data, transpose, distributed)
}
\arguments{
\item{data}{data set}

\item{transpose}{if we are transposing the data set}

\item{distributed}{if we are running distributed CoGAPS}
}
\value{
message displayed to screen
}
\description{
write start up message
}
