% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{stitchTogether}
\alias{stitchTogether}
\title{concatenate final results across subsets}
\usage{
stitchTogether(result, allParams)
}
\arguments{
\item{result}{list of CogapsResult object from all runs across subsets}

\item{allParams}{list of all CoGAPS parameters}
}
\value{
list with all CoGAPS output
}
\description{
concatenate final results across subsets
}
