% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{tf2ugFC}
\alias{tf2ugFC}
\title{Gene sets defined by transcription factors defined from TRANSFAC.}
\usage{
tf2ugFC
}
\description{
Gene sets defined by transcription factors defined from TRANSFAC.
}
