% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{toCSV}
\alias{toCSV}
\alias{toCSV,CogapsResult,character-method}
\title{save CoGAPS Result object as a set of csvs to directory 
see fromCSV}
\usage{
toCSV(object, save_location = ".")

\S4method{toCSV}{CogapsResult,character}(object, save_location = ".")
}
\arguments{
\item{object}{CogapsResult object}

\item{save_location}{directory to write to}
}
\value{
none
}
\description{
save as csv
}
