% Generated by roxygen2 (4.0.1): do not edit by hand
\name{binaryA}
\alias{binaryA}
\title{\code{binaryA} creates a binarized heatmap of the A matrix
in which the value is 1 if the value in Amean is greater than
threshold * Asd and 0 otherwise}
\usage{
binaryA(Amean, Asd, threshold = 3)
}
\arguments{
\item{Amean}{the mean estimate for the A matrix}

\item{Asd}{the standard deviations on Amean}

\item{threshold}{the number of standard deviations above zero
that an element of Amean must be to get a value of 1}
}
\description{
\code{binaryA} creates a binarized heatmap of the A matrix
in which the value is 1 if the value in Amean is greater than
threshold * Asd and 0 otherwise
}

