% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{calcCoGAPSStat}
\alias{calcCoGAPSStat}
\alias{calcCoGAPSStat,CogapsResult-method}
\title{calculate statistic on sets of measurements (genes) or samples}
\usage{
calcCoGAPSStat(
  object,
  sets = NULL,
  whichMatrix = "featureLoadings",
  numPerm = 1000,
  ...
)

\S4method{calcCoGAPSStat}{CogapsResult}(
  object,
  sets = NULL,
  whichMatrix = "featureLoadings",
  numPerm = 1000,
  ...
)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{sets}{list of sets of measurements/samples}

\item{whichMatrix}{either "featureLoadings" or "sampleFactors" indicating which matrix
to calculate the statistics for}

\item{numPerm}{number of permutations to use when calculatin p-value}

\item{...}{handles old arguments for backwards compatibility}
}
\value{
gene set statistics for each column of A
}
\description{
calculates a statistic to determine if a pattern is enriched in a
a particular set of measurements or samples.
}
