% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoGAPSStat.R
\name{calcCoGAPSStat}
\alias{calcCoGAPSStat}
\title{Calculate Gene Set Statistics}
\usage{
calcCoGAPSStat(Amean, Asd, GStoGenes, numPerm = 500)
}
\arguments{
\item{Amean}{A matrix mean values}

\item{Asd}{A matrix standard deviations}

\item{GStoGenes}{data.frame or list with gene sets}

\item{numPerm}{number of permutations for null}
}
\value{
gene set statistics for each column of A
}
\description{
Calculate Gene Set Statistics
}
\details{
calculates the gene set statistics for each
column of A using a Z-score from the elements of the A matrix,
the input gene set, and permutation tests
}
\examples{
data('SimpSim')
calcCoGAPSStat(SimpSim.result$Amean, SimpSim.result$Asd, GStoGenes=GSets,
numPerm=500)
}
