% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{calcGeneGSStat}
\alias{calcGeneGSStat}
\alias{calcGeneGSStat,CogapsResult-method}
\title{probability gene belongs in gene set}
\usage{
calcGeneGSStat(
  object,
  GStoGenes,
  numPerm,
  Pw = rep(1, ncol(object@featureLoadings)),
  nullGenes = FALSE
)

\S4method{calcGeneGSStat}{CogapsResult}(
  object,
  GStoGenes,
  numPerm,
  Pw = rep(1, ncol(object@featureLoadings)),
  nullGenes = FALSE
)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{GStoGenes}{data.frame or list with gene sets}

\item{numPerm}{number of permutations for null}

\item{Pw}{weight on genes}

\item{nullGenes}{logical indicating gene adjustment}
}
\value{
gene similiarity statistic
}
\description{
calculates the probability that a gene
listed in a gene set behaves like other genes in the set within
the given data set
}
