% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGeneGSStat.R
\name{calcGeneGSStat}
\alias{calcGeneGSStat}
\title{\code{calcGeneGSStat} calculates the probability that a gene
listed in a gene set behaves like other genes in the set within
the given data set}
\usage{
calcGeneGSStat(Amean, Asd, GSGenes, numPerm, Pw = rep(1, ncol(Amean)),
  nullGenes = F)
}
\arguments{
\item{Amean}{A matrix mean values}

\item{Asd}{A matrix standard deviations}

\item{GSGenes}{data.frame or list with gene sets}

\item{numPerm}{number of permutations for null}

\item{Pw}{weight on genes}

\item{nullGenes}{- logical indicating gene adjustment}
}
\description{
\code{calcGeneGSStat} calculates the probability that a gene
listed in a gene set behaves like other genes in the set within
the given data set
}

