% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{calcZ}
\alias{calcZ}
\alias{calcZ,CogapsResult-method}
\title{compute z-score matrix}
\usage{
calcZ(object, which = "feature")

\S4method{calcZ}{CogapsResult}(object, which = "feature")
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{which}{either "feature" or "sample" indicating which matrix to
calculate the z-score for}
}
\value{
matrix of z-scores
}
\description{
calculates the Z-score for each element based on input mean
and standard deviation matrices
}
\examples{
data(SimpSim)
result <- CoGAPS(SimpSim.data)
feature_zscore <- calcZ(result)
}
