% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{callInternalCoGAPS}
\alias{callInternalCoGAPS}
\title{make correct call to internal CoGAPS dispatch function}
\usage{
callInternalCoGAPS(index, sets, data, allParams, uncertainty, geneNames,
  sampleNames, fixedMatrix = NULL)
}
\arguments{
\item{index}{index for which subset to run on}

\item{sets}{list of all subsets}

\item{data}{data in a supported format}

\item{allParams}{list of all parameters}

\item{uncertainty}{uncertainty of data in the same format}

\item{geneNames}{names of all genes}

\item{sampleNames}{names of all samples}

\item{fixedMatrix}{matrix of matched patterns}
}
\value{
CogapsResult object
}
\description{
make correct call to internal CoGAPS dispatch function
}
\keyword{internal}
