% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\docType{methods}
\name{CoGAPS}
\alias{CoGAPS}
\alias{CoGAPS,character,CoGapsParams-method}
\alias{CoGAPS,matrix,CoGapsParams-method}
\title{CoGAPS Matrix Factorization Algorithm}
\usage{
CoGAPS(data, params = new("CoGapsParams"), uncertainty = NULL,
  checkpointFile = NULL, ...)

\S4method{CoGAPS}{matrix,CoGapsParams}(data, params = new("CoGapsParams"),
  uncertainty = NULL, checkpointFile = NULL, ...)

\S4method{CoGAPS}{character,CoGapsParams}(data, params = new("CoGapsParams"),
  uncertainty = NULL, checkpointFile = NULL, ...)
}
\arguments{
\item{data}{File name or R object (see details for supported types)}

\item{params}{CoGapsParams object}

\item{checkpointFile}{name of the checkpoint file}

\item{...}{keeps backwards compatibility with arguments from older versions}

\item{S}{uncertainty matrix (same supported types as data)}
}
\value{
CoGapsResult object
}
\description{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix
}
\details{
Currently, raw count matrices are the only supported R object. For
 file types CoGAPS supports csv, tsv, and mtx
}
\examples{
data(SimpSim)
params <- new("CoGapsParams")
result <- CoGAPS(SimpSim.D, params)
}

