% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\docType{methods}
\name{CoGAPS}
\alias{CoGAPS}
\title{CoGAPS Matrix Factorization Algorithm}
\usage{
CoGAPS(data, params = new("CogapsParams"), nThreads = NULL,
  messages = TRUE, outputFrequency = 1000, uncertainty = NULL,
  checkpointOutFile = "gaps_checkpoint.out", checkpointInterval = 1000,
  checkpointInFile = NULL, transposeData = FALSE, ...)
}
\arguments{
\item{data}{File name or R object (see details for supported types)}

\item{params}{CogapsParams object}

\item{uncertainty}{uncertainty matrix (same supported types as data)}

\item{checkpointInFile}{name of the checkpoint file}

\item{...}{keeps backwards compatibility with arguments from older versions}

\item{fixedMatrix}{data for fixing the values of either the A or P matrix;
used in conjuction with whichMatrixFixed (see CogapsParams)}
}
\value{
CogapsResult object
}
\description{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix
}
\details{
For file types CoGAPS supports csv, tsv, and mtx
}
\examples{
# Running from R object
data(GIST)
resultA <- CoGAPS(GIST.D)

# Running from file name
gist_path <- system.file("extdata/GIST.mtx", package="CoGAPS")
resultB <- CoGAPS(gist_path)

Setting Parameters
params <- new("CogapsParams")
params <- setParam(params, "nPatterns", 5)
resultC <- CoGAPS(GIST.D, params)
}

