% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\docType{methods}
\name{CoGAPS Matrix Factorization Algorithm}
\alias{CoGAPS}
\alias{CoGAPS Matrix Factorization Algorithm}
\alias{CoGAPS,SingleCellExperiment,CogapsParams-method}
\alias{CoGAPS,SummarizedExperiment,CogapsParams-method}
\alias{CoGAPS,character,CogapsParams-method}
\alias{CoGAPS,data.frame,CogapsParams-method}
\alias{CoGAPS,matrix,CogapsParams-method}
\title{CoGAPS}
\format{An object of class \code{NULL} of length 0.}
\usage{
\S4method{CoGAPS}{character,CogapsParams}(data, params = new("CogapsParams"),
  uncertainty = NULL, fixedMatrix = NULL, checkpointFile = NULL, ...)

\S4method{CoGAPS}{matrix,CogapsParams}(data, params = new("CogapsParams"),
  uncertainty = NULL, fixedMatrix = NULL, checkpointFile = NULL, ...)

\S4method{CoGAPS}{data.frame,CogapsParams}(data, params = new("CogapsParams"),
  uncertainty = NULL, fixedMatrix = NULL, checkpointFile = NULL, ...)

\S4method{CoGAPS}{SummarizedExperiment,CogapsParams}(data,
  params = new("CogapsParams"), uncertainty = NULL, fixedMatrix = NULL,
  checkpointFile = NULL, ...)

\S4method{CoGAPS}{SingleCellExperiment,CogapsParams}(data,
  params = new("CogapsParams"), uncertainty = NULL, fixedMatrix = NULL,
  checkpointFile = NULL, ...)
}
\arguments{
\item{data}{File name or R object (see details for supported types)}

\item{params}{CogapsParams object}

\item{uncertainty}{uncertainty matrix (same supported types as data)}

\item{fixedMatrix}{data for fixing the values of either the A or P matrix;
used in conjuction with whichMatrixFixed (see CogapsParams)}

\item{checkpointFile}{name of the checkpoint file}

\item{...}{keeps backwards compatibility with arguments from older versions}
}
\value{
CogapsResult object
}
\description{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix
}
\details{
Currently, raw count matrices are the only supported R object. For
 file types CoGAPS supports csv, tsv, and mtx
}
\examples{
# Running from R object
data(GIST)
resultA <- CoGAPS(GIST.D)
# Running from file name
gist_path <- system.file("extdata/GIST.mtx", package="CoGAPS")
resultB <- CoGAPS(gist_path)
Setting Parameters
params <- new("CogapsParams")
params <- setParam(params, "nPatterns", 5)
resultC <- CoGAPS(GIST.D, params)
}
\keyword{datasets}

