\name{CoGAPS-package}
\alias{CoGAPS-package}
\docType{package}
\title{CoGAPS: Coordinated Gene Analysis Pattern Sets
}
\description{
CoGAPS implements a Bayesian MCMC matrix factorization algorithm,
    GAPS, and links it to gene set statistic methods to infer
     biological process activity.  It can be used to perform
      sparse matrix factorization on any data, and when this 
      data represents biomolecules, to do gene set analysis.

\tabular{ll}{
Package: \tab CoGAPS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-23\cr
License: \tab LGPL\cr
} }

\author{
Maintainer: Elana J. Fertig <ejfertig@jhmi.edu>, Michael F. Ochs
    <ochsm@tcnj.edu>
}
\references{
Fertig EJ, Ding J, Favorov AV, Parmigiani G, Ochs MF.  
CoGAPS: an R/C++ package to identify patterns and biological
process activity in transcriptomic data. 
Bioinformatics. 2010 Nov 1;26(21):2792-3
}
