% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{CoGAPS}
\alias{CoGAPS}
\title{CoGAPS Matrix Factorization Algorithm}
\usage{
CoGAPS(D, S, ...)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix (std devs for chi-squared of Log Likelihood)}
}
\value{
list with A and P matrix estimates
}
\description{
CoGAPS Matrix Factorization Algorithm
}
\details{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix
}

