% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{CoGAPS}
\alias{CoGAPS}
\title{CoGAPS Matrix Factorization Algorithm}
\usage{
CoGAPS(D, S, nFactor = 7, nEquil = 1000, nSample = 1000,
  nOutputs = 1000, nSnapshots = 0, alphaA = 0.01, alphaP = 0.01,
  maxGibbmassA = 100, maxGibbmassP = 100, seed = -1, messages = TRUE,
  singleCellRNASeq = FALSE, whichMatrixFixed = "N",
  fixedPatterns = matrix(0), checkpointInterval = 0, ...)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix (std devs for chi-squared of Log Likelihood)}
}
\value{
list with A and P matrix estimates
}
\description{
CoGAPS Matrix Factorization Algorithm
}
\details{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix
}

