% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{CoGAPS}
\alias{CoGAPS}
\title{CoGAPS Matrix Factorization Algorithm}
\usage{
CoGAPS(data, params = new("CogapsParams"), nThreads = 1,
  messages = TRUE, outputFrequency = 500, uncertainty = NULL,
  checkpointOutFile = "gaps_checkpoint.out", checkpointInterval = 1000,
  checkpointInFile = NULL, transposeData = FALSE,
  subsetIndices = NULL, subsetDim = 0, BPPARAM = NULL,
  geneNames = NULL, sampleNames = NULL, fixedPatterns = NULL,
  whichMatrixFixed = "N", outputToFile = NULL, workerID = 1, ...)
}
\arguments{
\item{data}{File name or R object (see details for supported types)}

\item{params}{CogapsParams object}

\item{nThreads}{maximum number of threads to run on}

\item{messages}{T/F for displaying output}

\item{outputFrequency}{number of iterations between each output (set to 0 to
disable status updates, other output is controlled by @code messages)}

\item{uncertainty}{uncertainty matrix - either a matrix or a supported
file type}

\item{checkpointOutFile}{name of the checkpoint file to create}

\item{checkpointInterval}{number of iterations between each checkpoint (set
to 0 to disable checkpoints)}

\item{checkpointInFile}{if this is provided, CoGAPS runs from the checkpoint
contained in this file}

\item{transposeData}{T/F for transposing data while reading it in - useful
for data that is stored as samples x genes since CoGAPS requires data to be
genes x samples}

\item{subsetIndices}{set of indices to use from the data}

\item{subsetDim}{which dimension (1=rows, 2=cols) to subset}

\item{BPPARAM}{BiocParallel backend}

\item{geneNames}{vector of names of genes in data}

\item{sampleNames}{vector of names of samples in data}

\item{fixedPatterns}{fix either 'A' or 'P' matrix to these values}

\item{whichMatrixFixed}{either 'A' or 'P', indicating which matrix is fixed}

\item{outputToFile}{name of a file to save the output to, will create 4 files
of the form "filename_nPatterns_[Amean, Asd, Pmean, Psd].extension"}

\item{workerID}{if calling CoGAPS in parallel the worker ID can be specified,
only worker 1 prints output and each worker outputs when it finishes, this
is not neccesary when using the default parallel methods (i.e. distributed
CoGAPS) but only when the user is manually calling CoGAPS in parallel}

\item{...}{allows for overwriting parameters in params}
}
\value{
CogapsResult object
}
\description{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix
}
\details{
The supported R types are: matrix, data.frame, SummarizedExperiment,
SingleCellExperiment. The supported file types are csv, tsv, and mtx.
}
\examples{
# Running from R object
data(GIST)
resultA <- CoGAPS(GIST.data_frame, nIterations=100)

# Running from file name
gist_path <- system.file("extdata/GIST.mtx", package="CoGAPS")
resultB <- CoGAPS(gist_path, nIterations=100)

# Setting Parameters
params <- new("CogapsParams")
params <- setParam(params, "nPatterns", 5)
resultC <- CoGAPS(GIST.data_frame, params, nIterations=100)
}
