% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{CoGAPS}
\alias{CoGAPS}
\title{\code{CoGAPS} calls the C++ MCMC code through gapsRun and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix and then calls calcCoGAPSStat to estimate gene set
activity with nPerm set to 500}
\usage{
CoGAPS(data, unc, ABins = data.frame(), PBins = data.frame(), GStoGenes,
  nFactor = 7, simulation_id = "simulation", nEquil = 1000,
  nSample = 1000, nOutR = 1000, output_atomic = FALSE,
  fixedBinProbs = FALSE, fixedDomain = "N", sampleSnapshots = TRUE,
  numSnapshots = 100, plot = TRUE, nPerm = 500, alphaA = 0.01,
  nMaxA = 1e+05, max_gibbmass_paraA = 100, alphaP = 0.01, nMaxP = 1e+05,
  max_gibbmass_paraP = 100)
}
\arguments{
\item{data}{data matrix}

\item{unc}{uncertainty matrix (std devs for chi-squared of Log Likelihood)}

\item{ABins}{a matrix of same size as A which gives relative
probability of that element being non-zero}

\item{PBins}{a matrix of same size as P which gives relative
probability of that element being non-zero}

\item{GStoGenes}{data.frame or list with gene sets}

\item{nFactor}{number of patterns (basis vectors, metagenes)}

\item{simulation_id}{name to attach to atoms files if created}

\item{nEquil}{number of iterations for burn-in}

\item{nSample}{number of iterations for sampling}

\item{nOutR}{how often to print status into R by iterations}

\item{output_atomic}{whether to write atom files (large)}

\item{fixedBinProbs}{Boolean for using relative probabilities
given in Abins and Pbins}

\item{fixedDomain}{character to indicate whether A or P is
domain for relative probabilities}

\item{sampleSnapshots}{Boolean to indicate whether to capture
individual samples from Markov chain during sampling}

\item{numSnapshots}{the number of individual samples to capture}

\item{plot}{Boolean to indicate whether to produce output graphics}

\item{nPerm}{number of permutations in gene set test}

\item{alphaA}{sparsity parameter for A domain}

\item{nMaxA}{PRESENTLY UNUSED, future = limit number of atoms}

\item{max_gibbmass_paraA}{limit truncated normal to max size}

\item{alphaP}{sparsity parameter for P domain}

\item{nMaxP}{PRESENTLY UNUSED, future = limit number of atoms}

\item{max_gibbmass_paraP}{limit truncated normal to max size}
}
\description{
\code{CoGAPS} calls the C++ MCMC code through gapsRun and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix and then calls calcCoGAPSStat to estimate gene set
activity with nPerm set to 500
}


\details{
  CoGAPS first decomposes the data matrix using GAPS, \eqn{{\bf{D}}}, into a basis of underlying patterns and then determines the gene set activity in each of these patterns.

  The GAPS decomposition is achieved by finding amplitude and pattern matrices (\eqn{{\bf{A}}} and \eqn{{\bf{P}}}, respectively) for which \deqn{{\bf{D}} = {\bf{A}}{\bf{P}} + \Sigma,} where \eqn{\Sigma} is the matrix of uncertainties given by unc.  The matrices \eqn{\bf{A}} and \eqn{\bf{P}} are assumed to have the atomic prior described in Sibisi and Skilling (1997) and are found with MCMC sampling.

Then, the patterns identified in the columns of \eqn{\bf{P}} are linked to activity in each of the gene sets specified in GStoGenes using a novel z-score based statistic developed in Ochs et al. (2009).  Specifically, the z-score for pattern \eqn{p} and gene set \eqn{G_{i}} containing $G$ total genes is given by \deqn{Z_{i,p} = \frac{1}{G} \sum_{g in \mathcal{G_{i}}} {\frac{{\bf{A}_{gp}}}{Asd_{gp}}},}  
where \eqn{g} indexes the genes in the set and \eqn{Asd_{gp}} is the standard deviation of \eqn{{\bf{A}}_{gp}} obtained from MCMC sampling.  CoGAPS then uses the specified \code{nPerm} random sample tests to compute a consistent p value estimate from that z score.  Note that the data from Ochs et al. (2009) are provided with this package in GIST_TS_20084.RData and TFGSList.RData are also provided with this package for further validation.}

\value{
  A list containing:
  \item{meanChi2}{Value of \eqn{chi^2} for Amean and Pmean.}
  \item{D}{Data matrix \eqn{{\bf{D}}} input to factorization.}
  \item{Sigma}{uncertainty matrix (std devs for chi-squared of Log Likelihood)}
  \item{Amean}{Sampled mean value of the amplitude matrix \eqn{{\bf{A}}}.}
  \item{Asd}{Sampled standard deviation of the amplitude matrix \eqn{{\bf{A}}}.}  
  \item{Pmean}{Sampled mean value of the amplitude matrix \eqn{{\bf{P}}}.}
  \item{Psd}{Sampled standard deviation of the amplitude matrix \eqn{{\bf{P}}}.}
  \item{GSUpreg}{p-values for upregulation of each gene set in each pattern.}
  \item{GSDownreg}{p-values for downregulation of each gene set in each pattern.}
  \item{GSActEst}{p-values for activity of each gene set in each pattern.}
}

\examples{
\dontrun{
## Load data
nIter <- 5000

## Run GAPS matrix decomposition with gene set statistic
results <- CoGAPS(data=SimpSim.D, unc=SimpSim.S,
                  GStoGenes=GSets,
                  nFactor=3,
                  nEquil=nIter, nSample=nIter,
                  plot=FALSE)


## Plot the results
plotGAPS(results$Amean, results$Pmean, 'GSFigs')
}
}

\seealso{\code{\link{gapsRun}},\code{\link{calcCoGAPSStat}}}
\keyword{misc}
