% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{CoGapsFromCheckpoint}
\alias{CoGapsFromCheckpoint}
\title{Restart CoGAPS from Checkpoint File}
\usage{
CoGapsFromCheckpoint(D, S, path, checkpointFile = NA)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}

\item{path}{path to checkpoint file}

\item{checkpointFile}{name for future checkpooints made}
}
\value{
list with A and P matrix estimates
}
\description{
Restart CoGAPS from Checkpoint File
}
\details{
loads the state of a previous CoGAPS run from a file and
 continues the run from that point
}
