% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{CoGAPSFromFile}
\alias{CoGAPSFromFile}
\title{CoGAPS with file input for matrix}
\usage{
CoGAPSFromFile(D)
}
\arguments{
\item{D}{file path for data matrix}
}
\value{
list with A and P matrix estimates
}
\description{
CoGAPS with file input for matrix
}
\examples{
 file <- system.file("extdata/GIST.mtx", package="CoGAPS")
 CoGAPSFromFile(file)
}

