% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{compiledWithOpenMPSupport}
\alias{compiledWithOpenMPSupport}
\title{Check if compiler supported OpenMP}
\usage{
compiledWithOpenMPSupport()
}
\value{
true/false if OpenMP was supported
}
\description{
Check if compiler supported OpenMP
}
\examples{
CoGAPS::compiledWithOpenMPSupport()
}
