% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{corcut}
\alias{corcut}
\title{cluster patterns together}
\usage{
corcut(allPatterns, allParams)
}
\arguments{
\item{allPatterns}{matrix of all patterns across subsets}

\item{allParams}{list of all CoGAPS parameters}
}
\value{
patterns listed by which cluster they belong to
}
\description{
cluster patterns together
}
