% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{corrToMeanPattern}
\alias{corrToMeanPattern}
\title{calculate correlation of each pattern in a cluster to the cluster mean}
\usage{
corrToMeanPattern(cluster)
}
\value{
correlation of each pattern
}
\description{
calculate correlation of each pattern in a cluster to the cluster mean
}
\keyword{internal}
