% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGWCoGAPSSets.R
\name{createGWCoGAPSSets}
\alias{createGWCoGAPSSets}
\title{Create Gene Sets for GWCoGAPS}
\usage{
createGWCoGAPSSets(data = D, nSets = nSets,
  outRDA = "GenesInCoGAPSSets.Rda", keep = TRUE)
}
\arguments{
\item{data}{data matrix with unique rownames}

\item{nSets}{number of sets for parallelization}

\item{outRDA}{name of output file}

\item{keep}{logical indicating whether or not to save gene set list.}
}
\value{
list with randomly generated sets of genes from whole genome data
}
\description{
Create Gene Sets for GWCoGAPS
}
\details{
factors whole genome data into randomly generated sets for indexing
}
\examples{
\dontrun{
createGWCoGAPSSet(D,nSets=nSets)
}
}

