% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGWCoGAPSSets.R
\name{createGWCoGAPSSets}
\alias{createGWCoGAPSSets}
\title{Create Gene Sets for GWCoGAPS}
\usage{
createGWCoGAPSSets(D, S, nSets, simulationName)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}

\item{nSets}{number of sets to partition the data into}

\item{simulationName}{name used to identify files created by this simulation}
}
\value{
simulationName used to identify saved files
}
\description{
Create Gene Sets for GWCoGAPS
}
\details{
factors whole genome data into randomly generated sets for indexing
}
\examples{
data(SimpSim)
createGWCoGAPSSets(SimpSim.D, SimpSim.S, nSets=2, "example")
}
