% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{findConsensusMatrix}
\alias{findConsensusMatrix}
\title{find the consensus pattern matrix across all subsets}
\usage{
findConsensusMatrix(unmatchedPatterns, allParams)
}
\arguments{
\item{unmatchedPatterns}{list of all unmatched pattern matrices from initial
run of CoGAPS}

\item{allParams}{list of all CoGAPS parameters}
}
\value{
matrix of consensus patterns
}
\description{
find the consensus pattern matrix across all subsets
}
