% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{gapsMapRun}
\alias{gapsMapRun}
\title{Backwards Compatibility with v2}
\usage{
gapsMapRun(D, S, FP, ...)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}

\item{FP}{data.frame with rows giving fixed patterns for P}

\item{...}{v2 style parameters}
}
\value{
list with A and P matrix estimates
}
\description{
Backwards Compatibility with v2
}

