% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{gapsMapRun}
\alias{gapsMapRun}
\title{Backwards Compatibility with v2}
\usage{
gapsMapRun(D, S, FP, ABins = data.frame(), PBins = data.frame(),
  nFactor = 5, simulation_id = "simulation", nEquil = 1000,
  nSample = 1000, nOutR = 1000, output_atomic = FALSE,
  fixedMatrix = "P", fixedBinProbs = FALSE, fixedDomain = "N",
  sampleSnapshots = TRUE, numSnapshots = 100, alphaA = 0.01,
  nMaxA = 1e+05, max_gibbmass_paraA = 100, alphaP = 0.01, nMaxP = 1e+05,
  max_gibbmass_paraP = 100, seed = -1, messages = TRUE)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}

\item{FP}{data.frame with rows giving fixed patterns for P}

\item{ABins}{unused}

\item{PBins}{unused}

\item{nFactor}{number of patterns (basis vectors, metagenes), which must be
greater than or equal to the number of rows of FP}

\item{simulation_id}{unused}

\item{nEquil}{number of iterations for burn-in}

\item{nSample}{number of iterations for sampling}

\item{nOutR}{number of output messages}

\item{output_atomic}{unused}

\item{fixedMatrix}{unused}

\item{fixedBinProbs}{unused}

\item{fixedDomain}{unused}

\item{sampleSnapshots}{indicates if snapshots should be made}

\item{numSnapshots}{how many snapshots to take}

\item{alphaA}{sparsity parameter for A domain}

\item{nMaxA}{unused}

\item{max_gibbmass_paraA}{limit truncated normal to max size}

\item{alphaP}{sparsity parameter for P domain}

\item{nMaxP}{unused}

\item{max_gibbmass_paraP}{limit truncated normal to max size}

\item{seed}{a positive seed is used as-is, while any negative seed tells
the algorithm to pick a seed based on the current time}

\item{messages}{display progress messages}

\item{...}{v2 style parameters}
}
\value{
list with A and P matrix estimates
}
\description{
Backwards Compatibility with v2
}
\examples{
data(SimpSim)
nC <- ncol(SimpSim.D)
patterns <- matrix(1:nC/nC, nrow=1, ncol=nC)
result <- gapsMapRun(SimpSim.D, SimpSim.S, FP=patterns, nFactor=3)
}
