% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{gapsRun}
\alias{gapsRun}
\title{Backwards Compatibility with v2}
\usage{
gapsRun(D, S, ABins = data.frame(), PBins = data.frame(), nFactor = 7,
  simulation_id = "simulation", nEquil = 1000, nSample = 1000,
  nOutR = 1000, output_atomic = FALSE, fixedBinProbs = FALSE,
  fixedDomain = "N", sampleSnapshots = TRUE, numSnapshots = 100,
  alphaA = 0.01, nMaxA = 1e+05, max_gibbmass_paraA = 100, alphaP = 0.01,
  nMaxP = 1e+05, max_gibbmass_paraP = 100, seed = -1, messages = TRUE)
}
\arguments{
\item{D}{data matrix}

\item{S}{uncertainty matrix}
}
\value{
list with A and P matrix estimates
}
\description{
Backwards Compatibility with v2
}

