% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSeeds.R
\name{generateSeeds}
\alias{generateSeeds}
\title{Generate Seeds for Multiple Concurrent Runs}
\usage{
generateSeeds(chains = 2, seed = -1)
}
\arguments{
\item{chains}{number of seeds to generate (number of chains to run)}

\item{seed}{positive values are kept, negative values will be overwritten
by a seed generated from the current time}
}
\value{
vector of randomly generated seeds
}
\description{
Generate Seeds for Multiple Concurrent Runs
}
