% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSeeds.R
\name{generateSeeds}
\alias{generateSeeds}
\title{generateSeeds}
\usage{
generateSeeds(chains = 2, seed = -1)
}
\arguments{
\item{chains}{number of MCMC chains to be used}

\item{seed}{numeric indicating whether to generate seed from system clock. Default is -1.}
}
\value{
vector of randomly generated seeds for use with gapsRun, gapsMapRun, or GWCoGAPS
}
\description{
generateSeeds
}
\examples{
\dontrun{
generateSeeds(chains=2, seed=-1)
}


}

