% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{getAmplitudeMatrix}
\alias{getAmplitudeMatrix}
\alias{getAmplitudeMatrix,CogapsResult-method}
\alias{The}
\alias{method}
\alias{is}
\alias{same}
\alias{as}
\alias{getFeatureLoadings}
\alias{Alternate}
\alias{name}
\alias{for}
\title{return Amplitude  matrix from CogapsResult object}
\usage{
getAmplitudeMatrix(object)

\S4method{getAmplitudeMatrix}{CogapsResult}(object)
}
\arguments{
\item{object}{an object of type CogapsResult}
}
\value{
amplitude matrix
}
\description{
return Amplitude  matrix from CogapsResult object
}
\examples{
data(GIST)
amplitudeMatrix <- getAmplitudeMatrix(GIST.result)
}
