% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getGeneNames}
\alias{getGeneNames}
\title{extract gene names from data}
\usage{
getGeneNames(data, transpose)
}
\value{
vector of gene names
}
\description{
extract gene names from data
}
\keyword{internal}
