% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R, R/methods-CogapsParams.R
\docType{methods}
\name{getParam}
\alias{getParam}
\alias{getParam,CogapsParams-method}
\alias{getParam}
\title{get the value of a parameter}
\usage{
getParam(object, whichParam)

\S4method{getParam}{CogapsParams}(object, whichParam)
}
\arguments{
\item{object}{an object of type CogapsParams}

\item{whichParam}{a string with the name of the requested parameter}
}
\value{
the value of the parameter
}
\description{
get the value of a parameter
}
\examples{
 params <- new("CogapsParams")
 getParam(params, "seed")
}
