% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getSampleNames}
\alias{getSampleNames}
\title{extract sample names from data}
\usage{
getSampleNames(data, transpose)
}
\value{
vector of sample names
}
\description{
extract sample names from data
}
\keyword{internal}
