% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWCoGAPS.R
\name{GWCoGAPS}
\alias{GWCoGAPS}
\title{GWCoGAPS}
\usage{
GWCoGAPS(simulationName, nFactor, nCores = NA, cut = NA, minNS = NA, ...)
}
\arguments{
\item{nFactor}{number of patterns (basis vectors, metagenes), which must be
greater than or equal to the number of rows of FP}

\item{nCores}{number of cores for parallelization. If left to the default NA, nCores = nSets.}

\item{cut}{number of branches at which to cut dendrogram used in patternMatch4Parallel}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{...}{additional parameters to be fed into \code{gapsRun} and \code{gapsMapRun}}
}
\value{
list of A and P estimates
}
\description{
GWCoGAPS
}
\details{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix for whole genome data;
}
\examples{
# Load the sample data from CoGAPS
data(SimpSim)
# Run GWCoGAPS
GWCoGAPS(SimpSim.D, SimpSim.S, nFactor=3, nSets=2)
}
\seealso{
\code{\link{gapsRun}}, \code{\link{patternMatch4Parallel}}, and \code{\link{gapsMapRun}}
}

