% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWCoGAPS.R
\name{GWCoGAPS}
\alias{GWCoGAPS}
\title{GWCoGAPS}
\usage{
GWCoGAPS(simulationName, nFactor, nCores = NA, cut = NA, minNS = NA,
  manualMatch = FALSE, consensusPatterns = NULL,
  saveUnmatchedPatterns = FALSE, ...)
}
\arguments{
\item{simulationName}{name of this simulation}

\item{nFactor}{number of patterns (basis vectors, metagenes), which must be
greater than or equal to the number of rows of FP}

\item{nCores}{number of cores for parallelization. If left to the default NA, nCores = nSets.}

\item{cut}{number of branches at which to cut dendrogram used in patternMatch4Parallel}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{manualMatch}{logical indicating whether or not to stop after initial phase for manual pattern matching}

\item{consensusPatterns}{fixed pattern matrix to be used to ensure reciprocity of A weights accross sets}

\item{saveUnmatchedPatterns}{option to save intermediate results for each
set, before the pattern matching happens}

\item{...}{additional parameters to be fed into \code{gapsRun} and \code{gapsMapRun}}
}
\value{
list of A and P estimates
}
\description{
GWCoGAPS
}
\details{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix for whole genome data;
}
\examples{
data(SimpSim)
sim_name <- "example"
createGWCoGAPSSets(SimpSim.D, SimpSim.S, nSets=2, sim_name)
result <- GWCoGAPS(sim_name, nFactor=3, nEquil=200, nSample=200)
}
\seealso{
\code{\link{gapsRun}}, \code{\link{patternMatch4Parallel}}, and \code{\link{gapsMapRun}}
}
