% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWCoGAPS.R
\name{GWCoGapsFromCheckpoint}
\alias{GWCoGapsFromCheckpoint}
\title{Restart a GWCoGaps Run from Checkpoint}
\usage{
GWCoGapsFromCheckpoint(simulationName, nCores, cut = NA, minNS = NA, ...)
}
\arguments{
\item{simulationName}{name of this simulation}

\item{nCores}{number of cores for parallelization. If left to the default NA, nCores = nSets.}

\item{cut}{number of branches at which to cut dendrogram used in patternMatch4Parallel}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{...}{additional parameters to be fed into \code{gapsRun} and \code{gapsMapRun}}
}
\value{
list of A and P estimates
}
\description{
Restart a GWCoGaps Run from Checkpoint
}
\examples{
data(SimpSim)
sim_name <- "example"
createGWCoGAPSSets(SimpSim.D, SimpSim.S, nSets=2, sim_name)
trash <- GWCoGAPS(sim_name, nFactor=3, nEquil=200, nSample=200)
result <- GWCoGapsFromCheckpoint(sim_name, 2)
}
