% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R
\docType{methods}
\name{initialize,CogapsParams-method}
\alias{initialize,CogapsParams-method}
\title{constructor for CogapsParams}
\usage{
\S4method{initialize}{CogapsParams}(.Object, distributed = NULL, ...)
}
\arguments{
\item{.Object}{CogapsParams object}

\item{...}{initial values for slots}
}
\value{
initialized CogapsParams object
}
\description{
constructor for CogapsParams
}
