% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CoGapsParams.R
\docType{methods}
\name{initialize,CoGapsParams-method}
\alias{initialize,CoGapsParams-method}
\title{Constructor for CoGapsParams}
\usage{
\S4method{initialize}{CoGapsParams}(.Object, ...)
}
\value{
initialized CoGapsParams object
}
\description{
Constructor for CoGapsParams
}

