% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R
\docType{methods}
\name{initialize,CogapsParams-method}
\alias{initialize,CogapsParams-method}
\title{Constructor for CogapsParams}
\usage{
\S4method{initialize}{CogapsParams}(.Object, ...)
}
\value{
initialized CogapsParams object
}
\description{
Constructor for CogapsParams
}

