% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R
\docType{methods}
\name{initialize,CogapsResult-method}
\alias{initialize,CogapsResult-method}
\title{Constructor for CogapsResult}
\usage{
\S4method{initialize}{CogapsResult}(.Object, Amean, Pmean, Asd, Psd, seed,
  meanChiSq, diagnostics = NULL, ...)
}
\arguments{
\item{.Object}{CogapsResult object}

\item{Amean}{mean of sampled A matrices}

\item{Pmean}{mean of sampled P matrices}

\item{Asd}{std dev of sampled A matrices}

\item{Psd}{std dev of sampled P matrices}

\item{seed}{random seed used for the run}

\item{meanChiSq}{mean value of ChiSq statistic}

\item{diagnostics}{assorted diagnostic reports from the run}

\item{...}{initial values for slots}
}
\value{
initialized CogapsResult object
}
\description{
Constructor for CogapsResult
}
