% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R
\docType{methods}
\name{initialize,CogapsResult-method}
\alias{initialize,CogapsResult-method}
\title{Constructor for CogapsResult}
\usage{
\S4method{initialize}{CogapsResult}(.Object, Amean, Pmean, Asd, Psd, seed,
  meanChiSq, diagnostics = list(), ...)
}
\value{
initialized CogapsResult object
}
\description{
Constructor for CogapsResult
}
