% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R
\name{initialize,CogapsResult-method}
\alias{initialize,CogapsResult-method}
\title{Constructor for CogapsResult}
\usage{
\S4method{initialize}{CogapsResult}(
  .Object,
  Amean,
  Pmean,
  Asd,
  Psd,
  meanChiSq,
  geneNames,
  sampleNames,
  diagnostics = NULL,
  ...
)
}
\arguments{
\item{.Object}{CogapsResult object}

\item{Amean}{mean of sampled A matrices}

\item{Pmean}{mean of sampled P matrices}

\item{Asd}{std dev of sampled A matrices}

\item{Psd}{std dev of sampled P matrices}

\item{meanChiSq}{mean value of ChiSq statistic}

\item{geneNames}{names of genes in data}

\item{sampleNames}{names of samples in data}

\item{diagnostics}{assorted diagnostic reports from the run}

\item{...}{initial values for slots}
}
\value{
initialized CogapsResult object
}
\description{
Constructor for CogapsResult
}
