% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modsimdata}
\alias{modsimdata}
\title{Toy example to run CoGAPS on.}
\format{
'data.frame':	25 obs. of  20 variables.
}
\usage{
data(modsimdata)
}
\description{
\itemize{
  \item V1..V20. some variables, for example levels of gene expression
}
}
\keyword{datasets}
