% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R
\docType{methods}
\name{patternMarkers}
\alias{patternMarkers}
\alias{patternMarkers,CogapsResult-method}
\title{compute pattern markers statistic}
\usage{
patternMarkers(object, threshold = "all", lp = NA, full = FALSE)

\S4method{patternMarkers}{CogapsResult}(object, threshold = "all", lp = NA,
  full = FALSE)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{threshold}{the type of threshold to be used. The default "all" will
distribute genes into pattern with the lowest ranking. The "cut" thresholds
by the first gene to have a lower ranking, i.e. better fit to, a pattern.}

\item{lp}{a vector of weights for each pattern to be used for finding
markers. If NA markers for each pattern of the A matrix will be used.}

\item{full}{logical indicating whether to return the ranks of each gene for
each pattern}
}
\value{
By default a non-overlapping list of genes associated with each
\code{lp}. If \code{full=TRUE} a data.frame of genes rankings with a column
for each \code{lp} will also be returned.
}
\description{
calculate the most associated pattern for each gene
}

