% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{patternMarkers}
\alias{patternMarkers}
\alias{patternMarkers,CogapsResult-method}
\title{compute pattern markers statistic}
\usage{
patternMarkers(object, threshold = "all", lp = NA, axis = 1)

\S4method{patternMarkers}{CogapsResult}(object, threshold = "all", lp = NA, axis = 1)
}
\arguments{
\item{object}{an object of type CogapsResult}

\item{threshold}{the type of threshold to be used. The default "all" will
distribute genes into pattern with the lowest ranking. The "cut" thresholds
by the first gene to have a lower ranking, i.e. better fit to, a pattern.}

\item{lp}{a vector of weights for each pattern to be used for finding
markers. If NA markers for each pattern of the A matrix will be used.}

\item{axis}{either 1 or 2, specifying if pattern markers should be calculated using
the rows of the data (1) or the columns of the data (2)}
}
\value{
By default a non-overlapping list of genes associated with each
\code{lp}.
}
\description{
calculate the most associated pattern for each gene
}
\examples{
data(GIST)
pm <- patternMarkers(GIST.result)
}
