% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedCogaps.R
\name{patternMatch}
\alias{patternMatch}
\title{Match Patterns Across Multiple Runs}
\usage{
patternMatch(allPatterns, allParams)
}
\arguments{
\item{Atot}{a matrix containing the total by set estimates of Pmean output from \code{reOrderBySet}}

\item{nSets}{number of parallel sets used to generate \code{Atot}}

\item{cnt}{number of branches at which to cut dendrogram}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{maxNS}{maximum of individual set contributions a cluster must contain}

\item{ignore.NA}{logical indicating whether or not to ignore NAs from potential over dimensionalization. Default is FALSE.}

\item{bySet}{logical indicating whether to return list of matched set solutions from \code{Atot}}

\item{...}{additional parameters for \code{agnes}}
}
\value{
a matrix of consensus patterns by samples. If \code{bySet=TRUE} then a list of the set contributions to each
consensus pattern is also returned.
}
\description{
Match Patterns Across Multiple Runs
}
