% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotAtoms}
\alias{plotAtoms}
\title{\code{plotAtoms} a simple plot of the number of atoms
from one of the vectors returned with atom numbers}
\usage{
plotAtoms(gapsRes, type = "sampA")
}
\arguments{
\item{gapsRes}{the list resulting from applying GAPS}

\item{type}{the atoms to plot, values are \code{sampA}, \code{sampP} ,
\code{equilA}, or \code{equilP} to plot sampling or equilibration teop
atome numbers}
}
\description{
\code{plotAtoms} a simple plot of the number of atoms during the sampling period or equilibration periodfrom one of either A or P as specified in \code{type}.
}

