% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAtoms.R
\name{plotAtoms}
\alias{plotAtoms}
\title{\code{plotAtoms} a simple plot of the number of atoms
from one of the vectors returned with atom numbers}
\usage{
plotAtoms(gapsRes, type = "sampA")
}
\arguments{
\item{gapsRes}{the list resulting from applying GAPS}

\item{type}{the atoms to plot, values are "sampA", "sampP" ,
"equilA", or "equilP" to plot sampling or equilibration teop
atom numbers}
}
\description{
\code{plotAtoms} a simple plot of the number of atoms
from one of the vectors returned with atom numbers
}

