% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDiag}
\alias{plotDiag}
\title{\code{plotDiag} plots a series of diagnostic plots}
\usage{
plotDiag(gapsRes)
}
\arguments{
\item{gapsRes}{list returned by gapsRun, gapsMapRun, or CoGAPS}
}
\description{
\code{plotDiag} plots a series of diagnostic plots
}

