% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGAPS.R
\name{plotGAPS}
\alias{plotGAPS}
\title{\code{plotGAPS} plots the output A and P matrices as a
heatmap and line plot respectively}
\usage{
plotGAPS(A, P, outputPDF = "")
}
\arguments{
\item{A}{the mean A matrix}

\item{P}{the mean P matrix}

\item{outputPDF}{optional root name for PDF output, if
not specified, output goes to screen}
}
\description{
\code{plotGAPS} plots the output A and P matrices as a
heatmap and line plot respectively
}

