% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotP.R
\name{plotP}
\alias{plotP}
\title{\code{plotP} plots the P matrix in a line plot
with error bars}
\usage{
plotP(PMean_Mat, P_SD)
}
\arguments{
\item{PMean_Mat}{matrix of mean values of P}

\item{P_SD}{matrix of standard deviation values of P}
}
\description{
\code{plotP} plots the P matrix in a line plot
with error bars
}

