% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsResult.R, R/methods-CogapsResult.R
\docType{methods}
\name{plotPatternGeneSet}
\alias{plotPatternGeneSet}
\alias{plotPatternGeneSet,list,numeric,numeric-method}
\title{generate a barchart of most significant hallmark sets for a pattern}
\usage{
plotPatternGeneSet(patterngeneset, whichpattern = 1, padj_threshold = 0.05)

\S4method{plotPatternGeneSet}{list,numeric,numeric}(patterngeneset, whichpattern = 1, padj_threshold = 0.05)
}
\arguments{
\item{patterngeneset}{output from getPatternGeneSet}

\item{whichpattern}{which pattern to generate bar chart for}

\item{padj_threshold}{maximum adjusted p-value of gene sets rendered on the resulting plot}
}
\value{
image object of barchart
}
\description{
generate a barchart of most significant hallmark sets for a pattern
}
