% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPatternMarkers.R
\name{plotPatternMarkers}
\alias{plotPatternMarkers}
\title{plotPatternMarkers}
\usage{
plotPatternMarkers(data = NA, patternMarkers = PatternMarkers,
  patternPalette = NA, sampleNames = NA, samplePalette = NA,
  colDenogram = TRUE, heatmapCol = "bluered", scale = "row", ...)
}
\arguments{
\item{data}{<the dataset from which the patterns where generated>}

\item{patternMarkers}{<the list of genes generated from the patternMarkers function>}

\item{patternPalette}{<a vector indicating what color should be used for each pattern>}

\item{sampleNames}{<names of the samples to use for labeling >}

\item{samplePalette}{< a vector indicating what color should be used for each sample>}
}
\value{
heatmap of the \code{data} values for the \code{patternMarkers}
}
\description{
plotPatternMarkers
}
\examples{
\dontrun{
plotPatternMarkers(data=p,patternMarkers=PatternMarkers,patternPalette=NA,sampleNames=pd$sample,
samplePalette=pd$color,colDenogram=TRUE,heatmapCol="bluered", scale='row')
}
}
\seealso{
\code{\link{gplot}} and  \code{\link{heatmap.2}}
}

