% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSmoothPatterns.R
\name{plotSmoothPatterns}
\alias{plotSmoothPatterns}
\title{\code{plotSmoothPatterns} plots the output A and P matrices as a
heatmap and line plot respectively}
\usage{
plotSmoothPatterns(P, x = NULL, breaks = NULL, breakStyle = T,
  orderP = !all(is.null(x)), plotPTS = F, pointCol = "black",
  lineCol = "grey", add = F, ...)
}
\arguments{
\item{P}{the mean A matrix}

\item{x}{optional variables}

\item{breaks}{breaks in plots}

\item{breakStyle}{style of breaks}

\item{orderP}{whether to order patterns}

\item{plotPTS}{whether to plot points on lines}

\item{pointCol}{color of points}

\item{lineCol}{color of line}

\item{add}{logical specifying if bars should be added to an already existing plot; defaults to `FALSE'.}

\item{...}{arguments to be passed to/from other methods.  For the default method these can include further arguments (such as `axes', `asp' and `main') and graphical parameters (see `par') which are passed to `plot.window()', `title()' and `axis'.}
}
\description{
\code{plotSmoothPatterns} plots the output A and P matrices as a
heatmap and line plot respectively
}

