% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postFixed4Parallel.R
\name{postFixed4Parallel}
\alias{postFixed4Parallel}
\title{postFixed4Parallel}
\usage{
postFixed4Parallel(AP.fixed = NA, setPs = NA)
}
\arguments{
\item{AP.fixed}{output of parallel gapsMapRun calls with same FP}

\item{setPs}{data.frame with rows giving fixed patterns for P used as input for gapsMapRun}
}
\value{
list of two data.frames containing the A matrix estimates or their corresponding standard deviations
from output of parallel gapsMapRun
}
\description{
postFixed4Parallel
}

