% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructGene.R
\name{reconstructGene}
\alias{reconstructGene}
\title{reconstruct Gene}
\usage{
reconstructGene(A = AP$Amean, P = AP$Pmean, genes = NA)
}
\arguments{
\item{A}{A matrix estimates}

\item{P}{corresponding P matrix estimates}

\item{genes}{an indx of the gene or genes of interest. If \code{NA}, the default, all genes contained in A will be returned.}
}
\value{
the D' estimate of a gene or set of genes
}
\description{
reconstruct Gene
}

