% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderByPatternMatch.R
\name{reorderByPatternMatch}
\alias{reorderByPatternMatch}
\title{\code{reorderByPatternMatch} plots the output A and P matrices as a
heatmap and line plot respectively}
\usage{
reorderByPatternMatch(P, matchTo)
}
\arguments{
\item{P}{matrix to be matched}

\item{matchTo}{matrix to match P to}
}
\description{
\code{reorderByPatternMatch} plots the output A and P matrices as a
heatmap and line plot respectively
}

