% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reOrderBySet.R
\name{reOrderBySet}
\alias{reOrderBySet}
\title{reOrderBySet}
\usage{
reOrderBySet(AP, nFactor, nSets, match = "P")
}
\arguments{
\item{AP}{output of gapsRun in parallel}

\item{nFactor}{number of patterns}

\item{nSets}{number of sets}

\item{match}{which matrix to use for downstream matching. default is P}
}
\value{
a list containing the \code{nSets} sets solution for Amean under "A", Pmean under "P", and Asd under "Asd"
}
\description{
<restructures output of gapsRun into a list containing each sets solution for Amean, Pmean, and Asd>
}
