% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{Plot of Residuals}
\usage{
residuals(AMean_Mat, PMean_Mat, D, S)
}
\arguments{
\item{AMean_Mat}{matrix of mean values for A from GAPS}

\item{PMean_Mat}{matrix of mean values for P from GAPS}

\item{D}{original data matrix run through GAPS}

\item{S}{original standard deviation matrix run through GAPS}
}
\description{
Plot of Residuals
}
\details{
calculate residuals and produce heatmap
}
\examples{
# Load the simulated data
data('SimpSim')
# Load the outputs from gapsRun
data('results')
# Run residuals with the correct arguments
residuals(results$Amean,results$Pmean,SimpSim.D,SimpSim.S)
}

