% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCoGAPS.R
\name{scCoGAPS}
\alias{scCoGAPS}
\title{scCoGAPS}
\usage{
scCoGAPS(simulationName, nFactor, nCores = NA, cut = NA, minNS = NA,
  manualMatch = FALSE, consensusAs = NULL, ...)
}
\arguments{
\item{simulationName}{name of this simulation}

\item{nFactor}{number of patterns (basis vectors, metagenes), which must be
greater than or equal to the number of rows of FP}

\item{nCores}{number of cores for parallelization. If left to the default NA, nCores = nSets.}

\item{cut}{number of branches at which to cut dendrogram used in patternMatch4singleCell}

\item{minNS}{minimum of individual set contributions a cluster must contain}

\item{manualMatch}{logical indicating whether or not to stop after initial phase for manual pattern matching}

\item{consensusAs}{fixed pattern matrix to be used to ensure reciprocity of A weights accross sets}

\item{...}{additional parameters to be fed into \code{gapsRun} and \code{gapsMapRun}}
}
\value{
list of A and P estimates
}
\description{
scCoGAPS
}
\details{
calls the C++ MCMC code and performs Bayesian
matrix factorization returning the two matrices that reconstruct
the data matrix for whole genome data;
}
