% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CogapsParams.R, R/methods-CogapsParams.R
\docType{methods}
\name{setParam}
\alias{setParam}
\alias{setParam,CogapsParams-method}
\alias{setParam}
\title{set the value of a parameter}
\usage{
setParam(object, whichParam, value)

\S4method{setParam}{CogapsParams}(object, whichParam, value)
}
\arguments{
\item{object}{an object of type CogapsParams}

\item{whichParam}{a string with the name of the parameter to be changed}

\item{value}{the value to set the parameter to}
}
\value{
the modified params object
}
\description{
set the value of a parameter
}
\examples{
 params <- new("CogapsParams")
 params <- setParam(params, "seed", 123)
}
