% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{startupMessage}
\alias{startupMessage}
\title{write start up message}
\usage{
startupMessage(data, allParams)
}
\arguments{
\item{data}{data set}

\item{allParams}{list of all parameters}
}
\value{
message displayed to screen
}
\description{
write start up message
}
\keyword{internal}
