% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGAPS.R
\name{supported}
\alias{supported}
\title{Checks if file is supported}
\usage{
supported(file)
}
\arguments{
\item{file}{path to file}
}
\value{
TRUE if file is supported, FALSE if not
}
\description{
Checks if file is supported
}
