\name{tf2ugFC}
\docType{data}
\alias{tf2ugFC}
\title{Gene sets defined by transcription factors defined from TRANSFAC.}
\description{List of genes contained in gastrointestinal stromal tumor cell line measurements that are regulated by transcription factors in the TRANSFAC database.  Used for the gene set analysis in Ochs et al. (2009).
}
\usage{TFGSList}
\format{Data.frame containing genes (rows) regulated by each transcription factor (columns).}
\references{Ochs, M., Rink, L., Tarn, C., Mburu, S., Taguchi, T., Eisenberg, B., and Godwin, A. (2009). Detection of treatment-induced changes in signaling pathways in gastrointestinal stromal tumors using transcriptomic data. Cancer Res, 69(23), 9125-9132.}
\keyword{datasets}
